@@ -3,7 +3,8 @@ cmake_minimum_required(VERSION 3.15)
 project(
   PM
   VERSION 0.1.0
-  LANGUAGES CXX C)
+  LANGUAGES CXX C
+  DESCRIPTION "Branch-and-Price algorithm for parallel machine scheduling problems")
 
 #
 # Set project options
@@ -13,8 +14,9 @@ project(
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_SOURCE_DIR}/cmake
                       ${CMAKE_SOURCE_DIR})
 
-include(cmake/StandardSettings.cmake)
-include(cmake/Utils.cmake)
+include(StandardSettings)
+include(Utils)
+include(StaticAnalyzers)
 message(STATUS "Started CMake for ${PROJECT_NAME} v${PROJECT_VERSION}...\n")
 
 #
@@ -27,8 +29,8 @@ include(ConfigSafeGuards)
 # Enable package managers
 #
 
-include(cmake/Conan.cmake)
-include(cmake/Vcpkg.cmake)
+include(Conan)
+include(Vcpkg)
 
 #
 # code coverage
@@ -73,12 +75,12 @@ add_subdirectory(branch-and-bound)
 #
 
 if(${PROJECT_NAME}_BUILD_HEADERS_ONLY)
-  target_compile_features(${PROJECT_NAME} INTERFACE cxx_std_17)
+  target_compile_features(${PROJECT_NAME} INTERFACE cxx_std_20)
 else()
-  target_compile_features(${PROJECT_NAME} PUBLIC cxx_std_17)
+  target_compile_features(${PROJECT_NAME} PUBLIC cxx_std_20)
 endif()
 
-include(cmake/CompilerWarnings.cmake)
+include(CompilerWarnings)
 set_project_warnings(${PROJECT_NAME})
 
 verbose_message("Applied compiler warnings. Using standard ${CXX_STANDARD}.\n")
@@ -87,11 +89,20 @@ verbose_message("Applied compiler warnings. Using standard ${CXX_STANDARD}.\n")
 # -Wmissing-declarations -Wpointer-arith -Wcast-qual -Wcast-align
 # -Wwrite-strings)
 
-set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} -O5 -std=c++17 -m64")
+# RelWIthDebInfo Flags
+set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "${CMAKE_CXX_FLAGS_RELWITHDEBINFO} -O5 -std=c++20 -m64")
+set(CMAKE_C_FLAGS_RELWITHDEBINFO
+    "${CMAKE_C_FLAGS_RELWITHDEBINGFO} -std=c99 -m64 -D_XOPEN_SOURCE=500 -Wnested-externs -Wstrict-prototypes"
+)
+
+# Release Flags
+set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} -O5 -std=c++20 -m64")
 set(CMAKE_C_FLAGS_RELEASE
     "${CMAKE_C_FLAGS_RELEASE} -std=c99 -m64 -D_XOPEN_SOURCE=500 -Wnested-externs -Wstrict-prototypes"
 )
-set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -std=c++17 -O0 -m64")
+
+# Debug Flags
+set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -std=c++20 -O0 -m64")
 set(CMAKE_C_FLAGS_DEBUG
     "${CMAKE_C_FLAGS_DEBUG} -O0 -std=c99 -m64 -D_XOPEN_SOURCE=500 -Wnested-externs -Wstrict-prototypes"
 )
@@ -118,20 +129,21 @@ else()
       ${PROJECT_NAME}_LIB
       PUBLIC $<INSTALL_INTERFACE:include>
              $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-             ${CONAN_INCLUDE_DIRS_FMT}
              ${GUROBI_INCLUDE_DIR}
-             ${GLIB_INCLUDE_DIRS}
+             CONAN_PKG::fmt
+             CONAN_PKG::glib
+             CONAN_PKG::boost
              ORUtils::ORUtils
              branch-and-bound
       PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src)
-
-    conan_target_link_libraries(${PROJECT_NAME}_LIB)
 
     target_link_libraries(
       ${PROJECT_NAME}_LIB
       ${GUROBI_LIBRARY}
       ${GUROBI_CXX_LIBRARY}
-      ${GLIB_LIBRARIES}
+      CONAN_PKG::fmt
+      CONAN_PKG::glib
+      CONAN_PKG::boost
       ORUtils::ORUtils
       branch-and-bound
       -lm
