@@ -1,4 +1,6 @@
 #include "BranchBoundTree.hpp"
+#include <fmt/core.h>
+#include <boost/timer/timer.hpp>
 #include <memory>
 #include "BranchNode.hpp"
 #include "branch-and-bound/bfstree.h"
@@ -7,7 +9,6 @@
 #include "branch-and-bound/dfstree.h"
 #include "wctparms.h"
 #include "wctprivate.h"
-
 BranchBoundTree::BranchBoundTree(NodeData* root,
                                  int       _probType,
                                  bool      _isIntProb) {
@@ -27,13 +28,13 @@ BranchBoundTree::BranchBoundTree(NodeData* root,
             break;
     }
 
-    BranchNodeBase* node = new BranchNode(root, true);
+    auto node = std::make_unique<BranchNodeBase>(root, true);
     tree->setGlobalUB(double(root->opt_sol->tw));
     tree->setRetainStates(false);
-    node->setID(root->id);
-    node->setDepth(root->depth);
+    node->set_id(root->id);
+    node->set_depth(root->depth);
 
-    tree->processState(node, true);
+    tree->processState(std::move(node), true);
 }
 
 extern "C" {
