@@ -4,6 +4,8 @@
 #include <cstdio>
 #include <iterator>
 #include <limits>
+#include <memory>
+#include <span>
 #include "PricerSolverBase.hpp"
 #include "branch-and-bound/btree.h"
 #include "solver.h"
@@ -14,28 +16,50 @@
 BranchNodeBase::BranchNodeBase(NodeData* _pd, bool _isRoot)
     : State(_isRoot),
       pd(_pd) {
-    setDepth(pd->depth);
+    set_depth(pd->depth);
 }
 
 void BranchNodeBase::branch(BTree* bt) {
-    auto solver = pd->solver;
-    auto nb_jobs = pd->nb_jobs;
-    auto strong_branching =
+    auto* solver = pd->solver;
+    auto  nb_jobs = pd->nb_jobs;
+    auto  strong_branching =
         ((pd->opt_sol->tw - pd->LP_lower_bound) < (1.0 + IntegerTolerance));
 
     if (!strong_branching && dbg_lvl() > 0) {
         fmt::print("\nDOING STRONG BRANCHING...\n\n");
     }
 
-    fmt::print(
-        "BRANCHING NODE with branch_job = {} and middle_time = {} , less = "
-        "{}, "
-        "depth = {} with graph size {} and DWM LB = {}\n\n",
-        pd->branch_job, pd->completiontime, pd->less, depth,
-        solver->get_nb_vertices(), pd->LP_lower_bound);
+    // fmt::print(
+    //     "BRANCHING NODE with branch_job = {} and middle_time = {} , less = "
+    //     "{}, "
+    //     "depth = {} with graph size {}, DWM LB = {} and UB {}\n\n",
+    //     pd->branch_job, pd->completiontime, pd->less, get_depth(),
+    //     solver->get_nb_vertices(), pd->LP_lower_bound + pd->off,
+    //     pd->upper_bound + pd->off);
+
+    // fmt::print(
+    //     "{0:^10}{1:^10}|{2:^10}{3:^10}{10:^10}|{4:>10}{5:>10}{6:>10}|{7:>5}{8:>"
+    //     "5}|{9:^10}\n",
+    //     "Expl", "Unexpl", "Obj", "Depth", "Primal", "Dual", "Gap", "Job",
+    //     "Time", "Time", "Size");
+
+    if (pd->depth == 0 || bt->print_progress()) {
+        fmt::print(
+            "{0:>5}{1:>5}|{2:10.2f}{3:>10}{4:>10}|{5:10.2f}{6:10.2f}{7:10.2f}|{"
+            "8:>"
+            "5}{9:>5}|\n",
+            bt->tStats->get_nodes_explored(), bt->get_nb_nodes(),
+            pd->LP_lower_bound + pd->off, pd->depth, solver->get_nb_vertices(),
+            bt->getGlobalUB() + pd->off, bt->getGlobalLB() + pd->off, 0.0,
+            pd->branch_job, pd->completiontime);
+    }
 
     auto fathom_left = false;
     auto fathom_right = false;
+    if (bt->getGlobalUB() < pd->upper_bound) {
+        pd->upper_bound = static_cast<int>(bt->getGlobalUB());
+        solver->UB = bt->getGlobalUB();
+    }
 
     // calculate the fraction of each job finishing at each time in the
     // relaxation solution
@@ -44,7 +68,7 @@ void BranchNodeBase::branch(BTree* bt) {
         x_job_time[i].resize(pd->H_max + 1, 0.0);
     }
 
-    solver->calculate_job_time(x_job_time);
+    solver->calculate_job_time(&x_job_time);
 
     // initialize the order for evaluating the branching jobs
     std::vector<int> ord(nb_jobs);
@@ -61,11 +85,11 @@ void BranchNodeBase::branch(BTree* bt) {
                                             std::numeric_limits<double>::max());
 
     for (auto k = 0; k < nb_jobs; k++) {
-        auto i = ord[k];
-        auto prev = -1;
-        auto accum = 0.0;
-        auto dist_zero = 0.0;
-        auto job = static_cast<Job*>(g_ptr_array_index(solver->jobs, i));
+        auto  i = ord[k];
+        auto  prev = -1;
+        auto  accum = 0.0;
+        auto  dist_zero = 0.0;
+        auto* job = static_cast<Job*>(solver->jobs[i]);
         for (auto t = 0; t < pd->H_max + 1; t++) {
             accum += x_job_time[i][t];
             // avg_completion_time[i] +=
@@ -77,7 +101,7 @@ void BranchNodeBase::branch(BTree* bt) {
             // }
 
             if ((accum >= (1.0 - TargetBrTimeValue)) &&
-                (x_job_time[i][t] > 1e-3) && (prev != -1) &&
+                (x_job_time[i][t] > EPS) && (prev != -1) &&
                 (middle_time[i] == -1)) {
                 middle_time[i] = (t + job->processing_time + prev) / 2;
                 branch_scores[i] = double(middle_time[i]) * accum - dist_zero;
@@ -91,7 +115,7 @@ void BranchNodeBase::branch(BTree* bt) {
 
             dist_zero += double(t + job->processing_time) * x_job_time[i][t];
 
-            if (x_job_time[i][t] > 1e-3) {
+            if (x_job_time[i][t] > EPS) {
                 prev = t + job->processing_time;
             }
         }
@@ -107,20 +131,21 @@ void BranchNodeBase::branch(BTree* bt) {
         }
     }
 
-    auto            best_min_gain = 0.0;
-    auto            best_job = -1;
-    auto            best_time = 0;
-    BranchNodeBase* best_right = nullptr;
-    BranchNodeBase* best_left = nullptr;
+    auto      best_min_gain = 0.0;
+    auto      best_job = -1;
+    auto      best_time = 0;
+    std::span jobsarray{pd->jobarray->pdata, pd->jobarray->len};
+    std::unique_ptr<BranchNodeBase> best_right = nullptr;
+    std::unique_ptr<BranchNodeBase> best_left = nullptr;
     for (auto& it : best_cand) {
         auto i = it.job;
         if (i == -1) {
             continue;
         }
 
-        auto left_gain = 0.0;
-        auto right_gain = 0.0;
-        auto job = (Job*)g_ptr_array_index(pd->jobarray, i);
+        auto  left_gain = 0.0;
+        auto  right_gain = 0.0;
+        auto* job = static_cast<Job*>(jobsarray[i]);
         for (auto t = 0; t < pd->H_max + 1; t++) {
             if (t + job->processing_time <= middle_time[i]) {
                 left_gain += x_job_time[i][t];
@@ -138,9 +163,10 @@ void BranchNodeBase::branch(BTree* bt) {
                 best_time = middle_time[i];
             }
         } else {
-            auto left = new_node_data(pd);
-            auto left_solver = left->solver;
-            auto left_node_branch = new BranchNodeBase(left);
+            auto* left = new_node_data(pd);
+            auto* left_solver = left->solver;
+            auto  left_node_branch = std::make_unique<BranchNodeBase>(left);
+            // std::unique_ptr<BranchNodeBase>(new BranchNodeBase(left));
             left->solver->split_job_time(i, middle_time[i], true);
             left_node_branch->computeBounds(bt);
 
@@ -160,9 +186,9 @@ void BranchNodeBase::branch(BTree* bt) {
 
             // build the right node and solve its root LP only
 
-            auto right = new_node_data(pd);
-            auto right_solver = right->solver;
-            auto right_node_branch = new BranchNodeBase(right);
+            auto* right = new_node_data(pd);
+            auto* right_solver = right->solver;
+            auto  right_node_branch = std::make_unique<BranchNodeBase>(right);
 
             right_solver->split_job_time(i, middle_time[i], false);
             right_node_branch->computeBounds(bt);
@@ -188,24 +214,9 @@ void BranchNodeBase::branch(BTree* bt) {
                 best_min_gain = min_gain;
                 best_job = i;
                 best_time = middle_time[i];
-                if (best_right) {
-                    delete best_right;
-                }
-                best_right = right_node_branch;
 
-                if (best_left) {
-                    delete best_left;
-                }
-                best_left = left_node_branch;
-                right_node_branch = nullptr;
-                left_node_branch = nullptr;
-            }
-
-            if (left_node_branch) {
-                delete left_node_branch;
-            }
-            if (right_node_branch) {
-                delete right_node_branch;
+                best_right = std::move(right_node_branch);
+                best_left = std::move(left_node_branch);
             }
 
             if (fathom_left || fathom_right) {
@@ -215,16 +226,18 @@ void BranchNodeBase::branch(BTree* bt) {
     }
 
     if (best_job == -1) {
-        fprintf(stderr, "ERROR: no branching found!\n");
+        fmt::print(stderr, "ERROR: no branching found!\n");
         for (auto k = 0; k < nb_jobs; k++) {
-            auto j = ord[k];
-            auto job = (Job*)g_ptr_array_index(solver->jobs, j);
-            fprintf(stderr, "j=%d:", j);
-            for (int t = 0; t < pd->H_max; t++)
-                if (x_job_time[j][t] > 1e-12)
-                    fprintf(stderr, " (%d,%lg)", t + job->processing_time,
-                            x_job_time[j][t]);
-            fprintf(stderr, "\n");
+            auto  j = ord[k];
+            auto* job = static_cast<Job*>(solver->jobs[j]);
+            fmt::print(stderr, "j={}:", j);
+            for (int t = 0; t < pd->H_max; t++) {
+                if (x_job_time[j][t] > ERROR) {
+                    fmt::print(stderr, " ({},{})", t + job->processing_time,
+                               x_job_time[j][t]);
+                }
+            }
+            fmt::print(stderr, "\n");
         }
 
         exit(-1);
@@ -232,23 +245,23 @@ void BranchNodeBase::branch(BTree* bt) {
 
     /** Process the branching nodes insert them in the tree */
     if (strong_branching) {
-        auto left = new_node_data(pd);
-        auto left_solver = left->solver;
-        auto left_node_branch = new BranchNodeBase(left);
+        auto* left = new_node_data(pd);
+        auto* left_solver = left->solver;
+        auto  left_node_branch = std::make_unique<BranchNodeBase>(left);
         left_solver->split_job_time(best_job, best_time, false);
         left->branch_job = best_job;
         left->completiontime = best_time;
         left->less = 0;
-        bt->processState(left_node_branch);
+        bt->processState(std::move(left_node_branch));
 
-        auto right = new_node_data(pd);
-        auto right_solver = right->solver;
-        auto right_node_branch = new BranchNodeBase(right);
+        auto* right = new_node_data(pd);
+        auto* right_solver = right->solver;
+        auto  right_node_branch = std::make_unique<BranchNodeBase>(right);
         right_solver->split_job_time(best_job, best_time, true);
         right->branch_job = best_job;
         right->completiontime = best_time;
         right->less = 1;
-        bt->processState(right_node_branch);
+        bt->processState(std::move(right_node_branch));
         // auto job = (Job*)g_ptr_array_index(pd->solver->jobs, best_job);
         // fmt::print("NO STRONG BRANCHING {} {} {}\n\n", best_job, best_time,
         //            best_min_gain);
@@ -259,49 +272,48 @@ void BranchNodeBase::branch(BTree* bt) {
             best_time;
         best_left->pd->less = 0;
         best_right->pd->less = 1;
-        bt->processState(best_left);
-        bt->processState(best_right);
+        bt->processState(std::move(best_left));
+        bt->processState(std::move(best_right));
         bt->setStateComputesBounds(false);
     }
 
-    fmt::print("Number of Nodes in B&B tree = {}\n", bt->get_nb_nodes());
-    fmt::print("Number of nodes explored = {}\n",
-               bt->tStats->get_nodes_explored());
-    fmt::print("Branching choice: j = {}, t = {}, best_gain = {}\n", best_job,
-               best_time, best_min_gain);
+    bt->update_global_lb();
+
+    // fmt::print("Number of Nodes in B&B tree = {}\n", bt->get_nb_nodes());
+    // fmt::print("Number of nodes explored = {}\n",
+    //            bt->tStats->get_nodes_explored());
+    if (dbg_lvl() > 1) {
+        fmt::print("Branching choice: j = {}, t = {}, best_gain = {}\n",
+                   best_job, best_time, best_min_gain + pd->off);
+    }
 }
 
 void BranchNodeBase::computeBounds(BTree* bt) {
     build_rmp(pd);
     solve_relaxation(pd);
     compute_lower_bound(pd);
-    lowerBound = pd->LP_lower_bound;
-    objValue = pd->LP_lower_bound;
+    set_lb(pd->LP_lower_bound);
+    setObjValue(pd->LP_lower_bound);
 }
 
 void BranchNodeBase::assessDominance(State* otherState) {}
 
 bool BranchNodeBase::isTerminalState() {
-    auto solver = pd->solver;
+    auto* solver = pd->solver;
     return solver->get_is_integer_solution();
 }
 
 void BranchNodeBase::applyFinalPruningTests(BTree* bt) {}
 
-extern "C" {
-
-BranchNodeBase* new_branch_node(int _isRoot, NodeData* data) {
-    return new BranchNodeBase(data, _isRoot);
-};
-
-void delete_branch_node(BranchNodeBase* node) {
-    if (node) {
-        delete node;
-    }
+void BranchNodeBase::update_data(double upper_bound) {
+    auto* statistics = pd->stat;
+    statistics->global_upper_bound = static_cast<int>(upper_bound);
 }
 
+extern "C" {
+
 size_t call_getDepth(BranchNodeBase* state) {
-    return state->getDepth();
+    return state->get_depth();
 };
 int call_getDomClassID(BranchNodeBase* state) {
     return state->getDomClassID();
@@ -322,7 +334,7 @@ int call_getParentID(BranchNodeBase* state) {
     return state->getParentID();
 }
 void call_setID(BranchNodeBase* state, int i) {
-    state->setID(i);
+    state->set_id(i);
 }
 bool call_isDominated(BranchNodeBase* state) {
     return state->isDominated();
