@@ -7,10 +7,14 @@
 
 class BranchBoundTree {
    public:
-    BranchBoundTree(NodeData* data, int probType = 0, bool isIntProb = 1);
+    explicit BranchBoundTree(NodeData* data,
+                             int       probType = 0,
+                             bool      isIntProb = true);
     BranchBoundTree(BranchBoundTree&&) = default;
     BranchBoundTree& operator=(BranchBoundTree&&) = default;
-    ~BranchBoundTree(){};
+    BranchBoundTree& operator=(const BranchBoundTree&) = delete;
+    BranchBoundTree(const BranchBoundTree&) = delete;
+    ~BranchBoundTree() = default;
     BTree* get_ptr_tree() { return tree.get(); }
 
    private:
