@@ -4,24 +4,26 @@
 #include <fmt/core.h>
 #include <NodeBdd.hpp>
 #include <OptimalSolution.hpp>
+#include <array>
 #include "NodeBddEval.hpp"
+#include "math.h"
 
 template <typename T = double>
 class ForwardBddBase : public Eval<NodeBdd<T>, OptimalSolution<T>> {
-    OriginalModel<>* original_model;
-    double*          pi;
+    double* pi{};
 
    public:
-    ForwardBddBase()
-        : Eval<NodeBdd<T>, OptimalSolution<T>>(),
-          original_model(nullptr),
-          pi(nullptr) {}
+    ForwardBddBase() = default;
 
-    ForwardBddBase(const ForwardBddBase<T>& src) {}
+    ForwardBddBase<T>(const ForwardBddBase<T>& src) = default;
+    ForwardBddBase<T>(ForwardBddBase<T>&&) noexcept = default;
+    ForwardBddBase<T>& operator=(ForwardBddBase<T>&&) noexcept = default;
+    ForwardBddBase<T>& operator=(const ForwardBddBase<T>&) = default;
+    ~ForwardBddBase<T>() = default;
 
     void set_pi(double* _pi) { pi = _pi; }
 
-    const double* get_pi() const { return pi; }
+    [[nodiscard]] const double* get_pi() const { return pi; }
 
     virtual void initializenode(NodeBdd<T>& n) const = 0;
 
@@ -52,9 +54,13 @@ class ForwardBddBase : public Eval<NodeBdd<T>, OptimalSolution<T>> {
 template <typename T = double>
 class ForwardBddCycle : public ForwardBddBase<T> {
    public:
-    ForwardBddCycle() : ForwardBddBase<T>() {}
+    ForwardBddCycle() = default;
 
-    ForwardBddCycle(const ForwardBddCycle<T>& src) {}
+    ForwardBddCycle<T>(const ForwardBddCycle<T>& src) = default;
+    ForwardBddCycle<T>(ForwardBddCycle<T>&&) noexcept = default;
+    ForwardBddCycle<T>& operator=(const ForwardBddCycle<T>&) = default;
+    ForwardBddCycle<T>& operator=(ForwardBddCycle<T>&&) noexcept = default;
+    ~ForwardBddCycle<T>() = default;
 
     void initializenode(NodeBdd<T>& n) const override {
         if (n.get_weight() == 0) {
@@ -67,15 +73,15 @@ class ForwardBddCycle : public ForwardBddBase<T> {
     }
 
     void initializerootnode(NodeBdd<T>& n) const override {
-        n.forward_label[0].f = 0;
+        n.forward_label[0].get_f() = 0;
         n.forward_label[1].set_f(DBL_MAX / 2);
     }
 
     void evalNode(NodeBdd<T>& n) const override {
         Job* tmp_j = n.get_job();
         assert(tmp_j != nullptr);
-        double result;
-        bool   diff;
+        double result = NAN;
+        bool   diff = false;
 
         int   weight = n.get_weight();
         T     g;
@@ -167,12 +173,13 @@ class ForwardBddCycle : public ForwardBddBase<T> {
 template <typename T = double>
 class ForwardBddSimple : public ForwardBddBase<T> {
    public:
-    ForwardBddSimple()
-        : ForwardBddBase<T>(){
+    ForwardBddSimple<T>() = default;
 
-          };
-
-    ForwardBddSimple(const ForwardBddSimple<T>& src) {}
+    ForwardBddSimple(const ForwardBddSimple<T>&) = default;
+    ForwardBddSimple<T>& operator=(const ForwardBddSimple<T>&) = default;
+    ForwardBddSimple(ForwardBddSimple<T>&&) noexcept = default;
+    ForwardBddSimple<T>& operator=(ForwardBddSimple<T>&&) noexcept = default;
+    ~ForwardBddSimple<T>() = default;
 
     void initializenode(NodeBdd<T>& n) const override {
         if (n.get_weight() == 0) {
@@ -183,7 +190,7 @@ class ForwardBddSimple : public ForwardBddBase<T> {
     }
 
     void initializerootnode(NodeBdd<T>& n) const override {
-        n.forward_label[0].f = 0;
+        n.forward_label[0].get_f() = 0;
     }
 
     void evalNode(NodeBdd<T>& n) const override {
