@@ -1,10 +1,12 @@
 #ifndef NODE_BDD_REDUCER_HPP
 #define NODE_BDD_REDUCER_HPP
 
+#include <fmt/core.h>
 #include <cassert>
 #include <cmath>
 #include <cstddef>
 #include <ostream>
+#include <span>
 #include <stdexcept>
 
 #include <NodeBdd.hpp>
@@ -17,26 +19,26 @@
 
 template <typename T, bool BDD, bool ZDD>
 class DdReducer {
-    NodeTableEntity<T>&               input;
     TableHandler<T>                   oldDiagram;
+    NodeTableEntity<T>&               input;
     TableHandler<T>                   newDiagram;
     NodeTableEntity<T>&               output;
     std::vector<std::vector<NodeId>>  newIdTable;
     std::vector<std::vector<NodeId*>> rootPtr;
     int                               counter = 1;
 
-    struct ReducNodeInfo {
+    struct ReduceNodeInfo {
         NodeBdd<T> children;
         size_t     column;
 
-        size_t hash() const { return children.hash(); }
+        [[nodiscard]] size_t hash() const { return children.hash(); }
 
-        bool operator==(ReducNodeInfo const& o) const {
+        bool operator==(ReduceNodeInfo const& o) const {
             return children == o.children;
         }
 
-        friend std::ostream& operator<<(std::ostream&        os,
-                                        ReducNodeInfo const& o) {
+        friend std::ostream& operator<<(std::ostream&         os,
+                                        ReduceNodeInfo const& o) {
             return os << "(" << o.children << " -> " << o.column << ")";
         }
     };
@@ -45,14 +47,14 @@ class DdReducer {
 
    public:
     explicit DdReducer(TableHandler<T>& diagram, bool useMP = false)
-        : input(*diagram),
-          oldDiagram(diagram),
+        : oldDiagram(std::move(diagram)),
+          input(*oldDiagram),
           newDiagram(input.numRows()),
           output(*newDiagram),
           newIdTable(input.numRows()),
           rootPtr(input.numRows()),
           readyForSequentialReduction(false) {
-        diagram = newDiagram;
+        diagram = std::move(newDiagram);
 
         input.initTerminals();
         input.makeIndex(useMP);
@@ -69,18 +71,21 @@ class DdReducer {
      * in order to make lower-level index safe.
      */
     void makeReadyForSequentialReduction() {
-        if (readyForSequentialReduction)
+        if (readyForSequentialReduction) {
             return;
+        }
 
         for (int i = 2; i < input.numRows(); ++i) {
-            size_t const m = input[i].size();
-            T* const     tt = input[i].data();
+            // size_t const    m = input[i].size();
+            // std::span const tt{input[i].data(), input[i].size()};
+            // T* const     tt = input[i].data();
 
-            for (size_t j = 0; j < m; ++j) {
+            for (auto& it : input[i]) {
                 for (int b = 0; b < 2; ++b) {
-                    NodeId& f = tt[j].branch[b];
-                    if (f.row() == 0)
+                    NodeId& f = it.branch[b];
+                    if (f.row() == 0) {
                         continue;
+                    }
 
                     NodeId f0 = input.child(f, 0);
                     NodeId deletable = 0;
@@ -131,19 +136,21 @@ class DdReducer {
     void algorithmR(int i) {
         makeReadyForSequentialReduction();
         size_t const m = input[i].size();
-        T* const     tt = input[i].data();
+        // T* const     tt = input[i].data();
 
         std::vector<NodeId>& newId = newIdTable[i];
         newId.resize(m);
 
         for (size_t j = m - 1; j + 1 > 0; --j) {
-            NodeId& f0 = tt[j].branch[0];
-            NodeId& f1 = tt[j].branch[1];
+            NodeId& f0 = input[i][j].branch[0];
+            NodeId& f1 = input[i][j].branch[1];
 
-            if (f0.row() != 0)
+            if (f0.row() != 0) {
                 f0 = newIdTable[f0.row()][f0.col()];
-            if (f1.row() != 0)
+            }
+            if (f1.row() != 0) {
                 f1 = newIdTable[f1.row()][f1.col()];
+            }
 
             if (ZDD && f1 == 0) {
                 newId[j] = f0;
@@ -167,7 +174,7 @@ class DdReducer {
                 continue;
             }
 
-            NodeId& g0 = tt[j].branch[0];
+            NodeId& g0 = input[i][j].branch[0];
             assert(newId[j].row() == counter + 1);
             newId[j] = NodeId(counter, mm++, g0.hasEmpty());
         }
@@ -179,18 +186,18 @@ class DdReducer {
 
         if (mm > 0u) {
             output.initRow(counter, mm);
-            T* nt = output[counter].data();
+            std::span<T> nt{output[counter].data(), output[counter].size()};
 
             for (size_t j = 0; j < m; ++j) {
                 // NodeId const& f0 = tt[j].branch[0];
-                NodeId const& f1 = tt[j].branch[1];
+                NodeId const& f1 = input[i][j].branch[1];
 
                 if (ZDD && f1 == 0) {  // forwarded
                     assert(newId[j].row() < counter);
                 } else {
                     assert(newId[j].row() == counter);
                     size_t k = newId[j].col();
-                    nt[k] = tt[j];
+                    nt[k] = input[i][j];
                     nt[k].set_node_id_label(newId[j]);
                     if (nt[k].ptr_node_id != nullptr) {
                         *(nt[k].ptr_node_id) = newId[j];
@@ -201,8 +208,8 @@ class DdReducer {
             counter++;
         }
 
-        for (size_t k = 0; k < rootPtr[i].size(); ++k) {
-            NodeId& root = *rootPtr[i][k];
+        for (auto& k : rootPtr[i]) {
+            NodeId& root = *k;
             root = newId[root.col()];
         }
     }
@@ -213,9 +220,9 @@ class DdReducer {
      */
     void algorithmZdd(int i) {
         makeReadyForSequentialReduction();
-        size_t const m = input[i].size();
-        T* const     tt = input[i].data();
-        NodeId const mark(i, m);
+        size_t const       m = input[i].size();
+        std::span<T> const tt{input[i].data(), input[i].size()};
+        NodeId const       mark(i, m);
 
         auto& newId = newIdTable[i];
         newId.resize(m);
@@ -224,10 +231,12 @@ class DdReducer {
             NodeId& f0 = tt[j].branch[0];
             NodeId& f1 = tt[j].branch[1];
 
-            if (f0.row() != 0)
+            if (f0.row() != 0) {
                 f0 = newIdTable[f0.row()][f0.col()];
-            if (f1.row() != 0)
+            }
+            if (f1.row() != 0) {
                 f1 = newIdTable[f1.row()][f1.col()];
+            }
 
             if (ZDD && f1 == 0) {
                 newId[j] = f0;
@@ -256,8 +265,9 @@ class DdReducer {
         for (size_t j = 0; j < m; ++j) {
             NodeId const f(i, j);
             assert(newId[j].row() <= i + 1);
-            if (newId[j].row() <= i)
+            if (newId[j].row() <= i) {
                 continue;
+            }
 
             for (size_t k = j; k < m;) {  // for each g in f0-equivalent list
                 assert(j <= k);
@@ -291,7 +301,7 @@ class DdReducer {
 
         if (mm > 0u) {
             output.initRow(i, mm);
-            T* nt = output[i].data();
+            std::span<T> nt{output[i].data(), output[i].size()};
 
             for (size_t j = 0; j < m; ++j) {
                 NodeId const& f0 = tt[j].branch[0];
@@ -307,6 +317,7 @@ class DdReducer {
                     assert(newId[j].row() == i);
                     size_t k = newId[j].col();
                     nt[k] = tt[j];
+                    fmt::print("test {}\n", nt[k].coeff_list[1].size());
                     nt[k].set_node_id_label(newId[j]);
                 }
             }
@@ -314,8 +325,8 @@ class DdReducer {
             counter++;
         }
 
-        for (size_t k = 0; k < rootPtr[i].size(); ++k) {
-            NodeId& root = *rootPtr[i][k];
+        for (auto& k : rootPtr[i]) {
+            NodeId& root = *k;
             root = newId[root.col()];
         }
     }
@@ -348,8 +359,9 @@ class DdReducer {
                 for (int b = 1; b < 2; ++b) {
                     NodeId& ff = f.branch[b];
                     ff = newIdTable[ff.row()][ff.col()];
-                    if (ff != deletable)
+                    if (ff != deletable) {
                         del = false;
+                    }
                 }
 
                 if (del) {  // f is redundant
