@@ -24,6 +24,7 @@
 
 #pragma once
 
+#include <fmt/core.h>
 #include <algorithm>
 #include <cassert>
 #include <climits>
@@ -46,7 +47,7 @@
 template <typename T = NodeBdd<double>>
 class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
     TableHandler<T> diagram;  ///< The diagram structure.
-    NodeId          root_;    ///< Root node ID.
+    NodeId          root_{};  ///< Root node ID.
 
    public:
     /**
@@ -54,9 +55,10 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
      */
     DdStructure() : root_(0) {}
     DdStructure<T>(const DdStructure<T>&) = default;
-    DdStructure<T>(DdStructure<T>&&) = default;
-    DdStructure<T>& operator=(const DdStructure<T>&) = default;
-    DdStructure<T>& operator=(DdStructure<T>&&) = default;
+    DdStructure<T>(DdStructure<T>&&) noexcept = default;
+    DdStructure<T>& operator=(const DdStructure<T>&) = delete;
+    DdStructure<T>& operator=(DdStructure<T>&&) noexcept = default;
+    ~DdStructure<T>() = default;
 
     /**
      * Universal ZDD constructor.
@@ -125,11 +127,11 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
         if (n > 0) {
             for (int i = n; i > 0; --i) {
                 zs.subset(i);
-                diagram.derefLevel(i);
+                // diagram.derefLevel(i);
             }
         }
 
-        diagram = tmpTable;
+        diagram = std::move(tmpTable);
     }
 
    public:
@@ -143,7 +145,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
      * Gets the root node.
      * @return root node ID.
      */
-    NodeId root() const { return root_; }
+    [[nodiscard]] NodeId root() const { return root_; }
 
     /**
      * Gets a child node.
@@ -151,7 +153,9 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
      * @param b branch number.
      * @return child node ID.
      */
-    NodeId child(NodeId f, int b) const { return diagram->child(f, b); }
+    [[nodiscard]] NodeId child(NodeId f, int b) const {
+        return diagram->child(f, b);
+    }
 
     /**
      * Gets the diagram.
@@ -169,19 +173,19 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
      * Gets the level of the root node.
      * @return the level of root ZDD variable.
      */
-    int topLevel() const { return root_.row(); }
+    [[nodiscard]] int topLevel() const { return root_.row(); }
 
     /**
      * Gets the number of nonterminal nodes.
      * @return the number of nonterminal nodes.
      */
-    size_t size() const { return diagram->size(); }
+    [[nodiscard]] size_t size() const { return diagram->size(); }
 
     /**
      * Checks if DD is a 0-terminal only.
      * @return true if DD is a 0-terminal only.
      */
-    bool empty() const { return root_ == 0; }
+    [[nodiscard]] bool empty() const { return root_ == 0; }
 
     /**
      * Checks structural equivalence with another DD.
@@ -189,14 +193,18 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
      */
     bool operator==(DdStructure const& o) const {
         int n = root_.row();
-        if (n != o.root_.row())
+        if (n != o.root_.row()) {
             return false;
-        if (n == 0)
+        }
+        if (n == 0) {
             return root_ == o.root_;
-        if (root_ == o.root_ && &*diagram == &*o.diagram)
+        }
+        if (root_ == o.root_ && &*diagram == &*o.diagram) {
             return true;
-        if (size() > o.size())
+        }
+        if (size() > o.size()) {
             return o.operator==(*this);
+        }
 
         MyHashMap<InitializedNode, size_t> uniq;
         DataTable<NodeId>                  equiv(n + 1);
@@ -228,7 +236,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
                 }
 
                 size_t* p = uniq.getValue(node);
-                equiv[i][j] = NodeId(i, (p != 0) ? *p : m);
+                equiv[i][j] = NodeId(i, (p != nullptr) ? *p : m);
             }
         }
 
@@ -320,11 +328,11 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
     template <typename R>
     R evaluate_backward(Eval<T, R>& evaluator) {
         int   n = root_.row();
-        auto& work = *(getDiagram());
+        auto& work = *diagram;
         evaluator.set_table(&(*diagram));
 
         if (this->size() == 0) {
-            printf("empty DDstructure\n");
+            // fmt::print("empty DDstructure\n");
             R retval;
             return retval;
         }
@@ -347,7 +355,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
         evaluator.set_table(&(*diagram));
 
         if (this->size() == 0) {
-            printf("empty DDstructure\n");
+            // fmt::print("empty DDstructure\n");
             return;
         }
 
@@ -367,7 +375,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
         evaluator.set_table(&(*diagram));
 
         if (this->size() == 0) {
-            printf("empty DDstructure\n");
+            // fmt::print("empty DDstructure\n");
             R retval;
             return retval;
         }
@@ -404,7 +412,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
         evaluator.set_table(&(*diagram));
 
         if (this->size() == 0) {
-            printf("empty DDstructure\n");
+            // fmt::print("empty DDstructure\n");
             return;
         }
 
@@ -435,10 +443,10 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
      */
     class const_iterator {
         struct Selection {
-            NodeId node;
-            bool   val;
+            NodeId node{};
+            bool   val{false};
 
-            Selection() : val(false) {}
+            Selection() = default;
 
             Selection(NodeId _node, bool _val) : node(_node), val(_val) {}
 
@@ -458,8 +466,9 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
               cursor(begin ? -1 : -2),
               path(),
               itemset() {
-            if (begin)
+            if (begin) {
                 next(dd.root_);
+            }
         }
 
         const_iterator& operator++() {
@@ -497,8 +506,9 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
                     }
                 }
 
-                if (f == 1)
+                if (f == 1) {
                     break; /* found */
+                }
 
                 for (; cursor >= 0; --cursor) { /* up */
                     Selection& sel = path[cursor];
@@ -545,7 +555,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
     /**
      * Implements DdSpec.
      */
-    int getRoot(NodeId& f) const {
+    size_t getRoot(NodeId& f) const {
         f = root_;
         return (f == 1) ? -1 : f.row();
     }
@@ -553,7 +563,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
     /**
      * Implements DdSpec.
      */
-    int getChild(NodeId& f, [[maybe_unused]] int level, int value) const {
+    size_t getChild(NodeId& f, [[maybe_unused]] int level, int value) const {
         assert(level > 0 && level == f.row());
         assert(0 <= value && value < 2);
         f = child(f, value);
@@ -563,7 +573,7 @@ class DdStructure : public DdSpec<DdStructure<T>, NodeId> {
     /**
      * Implements DdSpec.
      */
-    size_t hashCode(NodeId const& f) const { return f.hash(); }
+    [[nodiscard]] size_t hashCode(NodeId const& f) const { return f.hash(); }
 
     /**
      * Dumps the node table in Sapporo ZDD format.
