@@ -1,13 +1,13 @@
 #ifndef NODE_BRANCH_ID_HPP
 #define NODE_BRANCH_ID_HPP
-#include <stddef.h>
+#include <cstddef>
 
 struct NodeBranchId {
     size_t col;
     int    row;
     int    val;
 
-    NodeBranchId() {}
+    NodeBranchId() = default;
 
     NodeBranchId(int _row, size_t _col, int _val)
         : col(_col),
