@@ -2,23 +2,20 @@
 #define OPTIMAL_SOLUTION_HPP
 #include <job.h>
 #include <iostream>
+#include <span>
 
 template <typename T = double>
 class OptimalSolution {
    public:
     T          obj;
-    int        cost;
-    int        C_max;
-    GPtrArray* jobs;
+    int        cost{};
+    int        C_max{};
+    GPtrArray* jobs{g_ptr_array_new()};
 
     /** Default constructor */
-    OptimalSolution() : obj(0), cost(0), C_max(0), jobs(g_ptr_array_new()) {}
+    OptimalSolution() : obj(0) {}
 
-    explicit OptimalSolution(T _obj)
-        : obj(_obj),
-          cost(0),
-          C_max(0),
-          jobs(g_ptr_array_new()) {}
+    explicit OptimalSolution(T _obj) : obj(_obj), jobs(g_ptr_array_new()) {}
 
     /** Copy constructor */
     OptimalSolution(const OptimalSolution& other)
@@ -101,10 +98,11 @@ class OptimalSolution {
     }
 
     void reverse_jobs() {
+        std::span aux{jobs->pdata, jobs->len};
         for (int low = 0, high = jobs->len - 1; low < high; low++, high--) {
-            auto temp = g_ptr_array_index(jobs, low);
-            g_ptr_array_index(jobs, low) = g_ptr_array_index(jobs, high);
-            g_ptr_array_index(jobs, high) = temp;
+            auto temp = aux[low];
+            aux[low] = aux[high];
+            aux[high] = temp;
         }
     }
 };
