@@ -1,7 +1,10 @@
 #include <glib.h>
 #include <interval.h>
 #include <solution.h>
+#include <boost/container_hash/hash_fwd.hpp>
 #include <boost/dynamic_bitset.hpp>
+#include <cstddef>
+#include <span>
 #include "NodeBddSpec.hpp"
 
 class conflict_state {
@@ -9,19 +12,30 @@ class conflict_state {
     boost::dynamic_bitset<> add;
     boost::dynamic_bitset<> remove;
 
-    conflict_state(){};
+    conflict_state(const conflict_state&) = default;
+    conflict_state(conflict_state&&) = default;
+    conflict_state& operator=(const conflict_state&) = default;
+    conflict_state& operator=(conflict_state&&) = default;
 
-    ~conflict_state(){};
+    conflict_state() = default;
+
+    ~conflict_state() = default;
 };
 
 class PricerConstruct : public DdSpec<PricerConstruct, int, 2> {
-    GPtrArray* pair_list;
-    int        nb_layers;
+    std::span<void*> aux_list;
+    int              nb_layers;
 
    public:
-    explicit PricerConstruct(GPtrArray* _pair_list) : pair_list(_pair_list) {
-        nb_layers = pair_list->len;
+    explicit PricerConstruct(GPtrArray* _pair_list)
+        : aux_list{_pair_list->pdata, _pair_list->len} {
+        nb_layers = _pair_list->len;
     };
+
+    PricerConstruct(const PricerConstruct&) = default;
+    PricerConstruct(PricerConstruct&&) = default;
+    PricerConstruct& operator=(const PricerConstruct&) = default;
+    PricerConstruct& operator=(PricerConstruct&&) = default;
 
     int getRoot(int& state) {
         state = 0;
@@ -29,19 +43,17 @@ class PricerConstruct : public DdSpec<PricerConstruct, int, 2> {
     };
 
     int getChild(int& state, int level, int value) const {
-        int layer = nb_layers - level;
-        int _j;
-        assert(0 <= layer && layer <= nb_layers - 1);
-        job_interval_pair* tmp_pair =
-            (job_interval_pair*)g_ptr_array_index(pair_list, layer);
-        interval* tmp_interval = tmp_pair->I;
-        Job*      tmp_j = tmp_pair->j;
+        auto layer = nb_layers - level;
+        // assert(0 <= layer && layer <= nb_layers - 1);
+        auto* tmp_pair = static_cast<job_interval_pair*>(aux_list[layer]);
+        auto* tmp_interval = static_cast<interval*>(tmp_pair->I);
+        auto* tmp_j = static_cast<Job*>(tmp_pair->j);
 
         if (value) {
             state = state + tmp_j->processing_time;
         }
 
-        _j = min_job(layer, state, value);
+        auto _j = min_job(layer, state, value);
 
         if (!(_j < nb_layers)) {
             if (state <= tmp_interval->b) {
@@ -53,19 +65,18 @@ class PricerConstruct : public DdSpec<PricerConstruct, int, 2> {
         return nb_layers - _j;
     }
 
-    ~PricerConstruct(){};
+    ~PricerConstruct() = default;
 
    private:
-    int min_job(int j, int state, int value) const {
-        int  val = nb_layers;
-        Job* tmp = ((job_interval_pair*)g_ptr_array_index(pair_list, j))->j;
+    [[nodiscard]] int min_job(int j, int state, int value) const {
+        int   val = nb_layers;
+        auto* tmp = static_cast<job_interval_pair*>(aux_list[j])->j;
 
         if (value) {
             for (int i = j + 1; i < nb_layers; ++i) {
-                job_interval_pair* tmp_pair =
-                    (job_interval_pair*)g_ptr_array_index(pair_list, i);
-                interval* tmp_interval = tmp_pair->I;
-                Job*      tmp_j = tmp_pair->j;
+                auto* tmp_pair = static_cast<job_interval_pair*>(aux_list[i]);
+                auto* tmp_interval = static_cast<interval*>(tmp_pair->I);
+                auto* tmp_j = static_cast<Job*>(tmp_pair->j);
 
                 if (state + tmp_j->processing_time > tmp_interval->a &&
                     state + tmp_j->processing_time <= tmp_interval->b) {
@@ -78,10 +89,9 @@ class PricerConstruct : public DdSpec<PricerConstruct, int, 2> {
             }
         } else {
             for (int i = j + 1; i < nb_layers; ++i) {
-                job_interval_pair* tmp_pair =
-                    (job_interval_pair*)g_ptr_array_index(pair_list, i);
-                interval* tmp_interval = tmp_pair->I;
-                Job*      tmp_j = tmp_pair->j;
+                auto* tmp_pair = static_cast<job_interval_pair*>(aux_list[i]);
+                auto* tmp_interval = static_cast<interval*>(tmp_pair->I);
+                auto* tmp_j = static_cast<Job*>(tmp_pair->j);
 
                 if (state + tmp_j->processing_time > tmp_interval->a &&
                     state + tmp_j->processing_time <= tmp_interval->b) {
@@ -102,18 +112,23 @@ class PricerConstruct : public DdSpec<PricerConstruct, int, 2> {
 };
 
 class PricerConstructTI : public DdSpec<PricerConstructTI, int, 2> {
-    GPtrArray* pair_list;
-    int*       take_job;
-    int        Hmax;
-    int        nb_layers;
+    std::span<void*> aux_list;
+    // GPtrArray*       pair_list;
+    int* take_job;
+    int  Hmax;
+    int  nb_layers;
 
    public:
     explicit PricerConstructTI(GPtrArray* _pair_list, int* _take_job, int _Hmax)
-        : pair_list(_pair_list),
+        : aux_list{_pair_list->pdata, _pair_list->len},
           take_job(_take_job),
-          Hmax(_Hmax) {
-        nb_layers = pair_list->len;
-    };
+          Hmax(_Hmax),
+          nb_layers(_pair_list->len){};
+
+    PricerConstructTI(const PricerConstructTI&) = default;
+    PricerConstructTI(PricerConstructTI&&) = default;
+    PricerConstructTI& operator=(const PricerConstructTI&) = default;
+    PricerConstructTI& operator=(PricerConstructTI&&) = default;
 
     int getRoot(int& state) {
         state = 0;
@@ -121,13 +136,11 @@ class PricerConstructTI : public DdSpec<PricerConstructTI, int, 2> {
     };
 
     int getChild(int& state, int level, int value) const {
-        int layer = nb_layers - level;
-        int _j;
-        assert(0 <= layer && layer <= nb_layers - 1);
-        job_interval_pair* tmp_pair =
-            (job_interval_pair*)g_ptr_array_index(pair_list, layer);
-        interval* tmp_interval = tmp_pair->I;
-        Job*      tmp_j = tmp_pair->j;
+        auto layer = nb_layers - level;
+        // assert(0 <= layer && layer <= nb_layers - 1);
+        auto* tmp_pair = static_cast<job_interval_pair*>(aux_list[layer]);
+        auto* tmp_interval = static_cast<interval*>(tmp_pair->I);
+        auto* tmp_j = static_cast<Job*>(tmp_pair->j);
 
         if (level - 1 == 0 && value) {
             return (state + tmp_j->processing_time <= tmp_interval->b) ? -1 : 0;
@@ -139,7 +152,7 @@ class PricerConstructTI : public DdSpec<PricerConstructTI, int, 2> {
             state = state + tmp_j->processing_time;
         }
 
-        _j = min_job(layer, state, value);
+        auto _j = min_job(layer, state, value);
 
         if (!(_j < nb_layers)) {
             if (state <= tmp_interval->b) {
@@ -151,19 +164,19 @@ class PricerConstructTI : public DdSpec<PricerConstructTI, int, 2> {
         return nb_layers - _j;
     }
 
-    ~PricerConstructTI(){};
+    ~PricerConstructTI() = default;
 
    private:
-    int min_job(int j, int state, int value) const {
+    [[nodiscard]] int min_job(int j, int state, int value) const {
         int                val = nb_layers;
-        job_interval_pair* tmp_pair;
-        interval*          tmp_interval;
-        Job*               tmp_j;
-        Job* tmp = ((job_interval_pair*)g_ptr_array_index(pair_list, j))->j;
+        job_interval_pair* tmp_pair = nullptr;
+        interval*          tmp_interval = nullptr;
+        Job*               tmp_j = nullptr;
+        auto* tmp = (static_cast<job_interval_pair*>(aux_list[j]))->j;
 
         if (value) {
             for (int i = j + 1; i < nb_layers; ++i) {
-                tmp_pair = (job_interval_pair*)g_ptr_array_index(pair_list, i);
+                tmp_pair = static_cast<job_interval_pair*>(aux_list[i]);
                 tmp_interval = tmp_pair->I;
                 tmp_j = tmp_pair->j;
 
@@ -181,7 +194,7 @@ class PricerConstructTI : public DdSpec<PricerConstructTI, int, 2> {
             }
         } else {
             for (int i = j + 1; i < nb_layers; ++i) {
-                tmp_pair = (job_interval_pair*)g_ptr_array_index(pair_list, i);
+                tmp_pair = static_cast<job_interval_pair*>(aux_list[i]);
                 tmp_interval = tmp_pair->I;
                 tmp_j = tmp_pair->j;
 
@@ -227,12 +240,15 @@ class ConflictConstraints
     }
 
    public:
-    ConflictConstraints(int  _nb_jobs,
-                        int* elist_same,
-                        int  edge_count_same,
-                        int* elist_differ,
-                        int  edge_count_differ)
+    ConflictConstraints(int    _nb_jobs,
+                        int*   elist_same,
+                        size_t edge_count_same,
+                        int*   elist_differ,
+                        size_t edge_count_differ)
         : nb_jobs(_nb_jobs) {
+        std::span aux_same{elist_same, edge_count_same};
+        std::span aux_diff{elist_differ, edge_count_differ};
+
         differsets.resize(_nb_jobs);
         samesets.resize(_nb_jobs);
 
@@ -242,15 +258,19 @@ class ConflictConstraints
         }
 
         for (int i = 0; i < edge_count_same; ++i) {
-            samesets[elist_same[2 * i]][elist_same[2 * i + 1]] = 1;
+            samesets[aux_same[2 * i]][aux_same[2 * i + 1]] = true;
         }
 
         for (int i = 0; i < edge_count_differ; ++i) {
-            differsets[elist_differ[2 * i]][elist_differ[2 * i + 1]] = 1;
+            differsets[aux_diff[2 * i]][aux_diff[2 * i + 1]] = true;
         }
     };
 
-    ~ConflictConstraints() {}
+    ~ConflictConstraints() = default;
+    ConflictConstraints(const ConflictConstraints&) = default;
+    ConflictConstraints(ConflictConstraints&&) = default;
+    ConflictConstraints& operator=(ConflictConstraints&&) = default;
+    ConflictConstraints& operator=(const ConflictConstraints&) = default;
 
     int getRoot(conflict_state& state) const {
         state.add.resize(nb_jobs);
@@ -260,7 +280,7 @@ class ConflictConstraints
 
     int getChild(conflict_state& state, int level, int take) {
         int job = nb_jobs - level;
-        int _j;
+        int _j = 0;
         assert(0 <= job && job <= nb_jobs - 1);
 
         if (samesets[job].intersects(differsets[job])) {
@@ -288,7 +308,7 @@ class ConflictConstraints
             state.remove |= samesets[job];
         }
 
-        _j = min_job(job, state);
+        _j = min_job(job, &state);
 
         if (_j == nb_jobs && take) {
             return (!state.remove[job]) ? -1 : 0;
@@ -300,8 +320,8 @@ class ConflictConstraints
         return nb_jobs - _j;
     }
 
-    bool equalTo(conflict_state const& state1,
-                 conflict_state const& state2) const {
+    [[nodiscard]] bool equalTo(conflict_state const& state1,
+                               conflict_state const& state2) const {
         if (state2.add != state1.add) {
             return false;
         }
@@ -313,30 +333,32 @@ class ConflictConstraints
         return true;
     }
 
-    size_t hashCode(conflict_state const& state) const {
+    [[nodiscard]] size_t hashCode(conflict_state const& state) const {
         size_t val = 0;
         size_t it = state.add.find_first();
 
         while (it != boost::dynamic_bitset<>::npos) {
-            val += 1213657 * it;
+            // val += 1213657 * it;
+            boost::hash_combine(val, it);
             it = state.add.find_next(it);
         }
 
         it = state.remove.find_first();
 
         while (it != boost::dynamic_bitset<>::npos) {
-            val += 487239 * it;
+            // val += 487239 * it;
+            boost::hash_combine(val, it);
             it = state.remove.find_next(it);
         }
 
         return val;
     }
 
-    int min_job(int j, conflict_state& state) const {
-        int i, val = nb_jobs;
+    int min_job(int j, conflict_state* state) const {
+        int val = nb_jobs;
 
-        for (i = j + 1; i < nb_jobs; ++i) {
-            if (!state.remove[i]) {
+        for (int i = j + 1; i < nb_jobs; ++i) {
+            if (!state->remove[i]) {
                 val = i;
                 break;
             }
@@ -347,20 +369,25 @@ class ConflictConstraints
 };
 
 class scheduling : public DdSpec<scheduling, int, 2> {
-    Job*       job;
-    GPtrArray* list_layers;
-    int        nb_layers;
-    int        order;
+    Job*             job;
+    GPtrArray*       list_layers;
+    std::span<void*> aux_list;
+    int              nb_layers;
+    int              order;
 
    public:
     scheduling(Job* _job, GPtrArray* _list_layers, int _order)
         : job(_job),
           list_layers(_list_layers),
-          order(_order) {
-        nb_layers = list_layers->len;
-    };
+          aux_list{_list_layers->pdata, _list_layers->len},
+          nb_layers(_list_layers->len),
+          order(_order){};
 
-    ~scheduling() {}
+    scheduling(const scheduling&) = default;
+    scheduling(scheduling&&) = default;
+    scheduling& operator=(const scheduling&) = default;
+    scheduling& operator=(scheduling&&) = default;
+    ~scheduling() = default;
 
     int getRoot(int& state) const {
         state = 0;
@@ -368,11 +395,10 @@ class scheduling : public DdSpec<scheduling, int, 2> {
     }
 
     int getChild(int& state, int level, int take) {
-        int                j = nb_layers - level;
-        job_interval_pair* tmp =
-            (job_interval_pair*)g_ptr_array_index(list_layers, j);
-        Job* tmp_j = tmp->j;
-        assert(0 <= j && j <= nb_layers - 1);
+        auto j = nb_layers - level;
+        // assert(0 <= j && j <= nb_layers - 1);
+        auto* tmp = static_cast<job_interval_pair*>(aux_list[j]);
+        auto* tmp_j = static_cast<Job*>(tmp->j);
 
         if (level - 1 == 0 && take) {
             if (tmp_j != job) {
