@@ -28,6 +28,17 @@ typedef enum {
 } NodeDataStatus;
 
 /**
+ *  CONSTANTS NODEDATA STRUCTURE
+ *
+ */
+
+#define NB_CUTS (2000)
+#define NB_CG_ITERATIONS (1000000)
+#define CLEANUP_ITERATION (30)
+#define EPS (1e-6)
+#define EPS_BOUND (1e-9)
+
+/**
  * problem data
  */
 typedef struct _Problem Problem;
@@ -40,18 +51,16 @@ typedef struct _NodeData NodeData;
 typedef struct BranchNodeBase  BranchNode;
 typedef struct BranchBoundTree BranchBoundTree;
 
-BranchNode* new_branch_node(int _isRoot, NodeData* data);
-void        delete_branch_node(BranchNode* node);
-size_t      call_getDepth(BranchNode* state);
-int         call_getDomClassID(BranchNode* state);
-double      call_getObjValue(BranchNode* state);
-double      call_getLB(BranchNode* state);
-double      call_getUB(BranchNode* state);
-int         call_getID(BranchNode* state);
-int         call_getParentID(BranchNode* state);
-void        call_setID(BranchNode* state, int i);
-int         isDominated(BranchNode* state);
-int         wasProcessed(BranchNode* state);
+size_t call_getDepth(BranchNode* state);
+int    call_getDomClassID(BranchNode* state);
+double call_getObjValue(BranchNode* state);
+double call_getLB(BranchNode* state);
+double call_getUB(BranchNode* state);
+int    call_getID(BranchNode* state);
+int    call_getParentID(BranchNode* state);
+void   call_setID(BranchNode* state, int i);
+int    isDominated(BranchNode* state);
+int    wasProcessed(BranchNode* state);
 
 BranchBoundTree* new_branch_bound_tree(NodeData* data,
                                        int       _probtype,
