@@ -2,6 +2,7 @@
 #include <fmt/core.h>
 #include <algorithm>
 #include <boost/concept_archetype.hpp>
+#include <cassert>
 #include <complex>
 #include <cstddef>
 #include <cstdio>
@@ -29,38 +30,28 @@
 #include "util.h"
 #include "wctprivate.h"
 
-using namespace std;
+using std::list;
+using std::vector;
 
 PricerSolverBdd::PricerSolverBdd(GPtrArray*  _jobs,
                                  int         _num_machines,
                                  GPtrArray*  _ordered_jobs,
-                                 const char* p_name,
-                                 int         _Hmax,
+                                 const char* _p_name,
+                                 int         _hmax,
                                  int*        _take_jobs,
-                                 double      _UB)
-    : PricerSolverBase(_jobs, _num_machines, p_name, _UB),
-      size_graph(0),
-      nb_removed_edges(0),
-      nb_removed_nodes(0),
-      ordered_jobs(_ordered_jobs),
-      original_model(reformulation_model),
-      H_min(0),
-      H_max(_Hmax)
+                                 double      _ub)
+    : PricerSolverBase(_jobs, _num_machines, _p_name, _ub),
+      decision_diagram{PricerConstruct(_ordered_jobs)},
+      size_graph{0},
+      ordered_jobs{_ordered_jobs},
+      original_model{reformulation_model},
+      H_min{},
+      H_max{_hmax}
 
 {
-    /**
-     * Construction of decision diagram
-     */
-    if (_take_jobs) {
-        PricerConstructTI ps(ordered_jobs, _take_jobs, _Hmax);
-        decision_diagram = std::make_unique<DdStructure<>>(ps);
-    } else {
-        PricerConstruct ps(ordered_jobs);
-        decision_diagram = std::make_unique<DdStructure<>>(ps);
-    }
     remove_layers_init();
-    decision_diagram->compressBdd();
-    size_graph = decision_diagram->size();
+    decision_diagram.compressBdd();
+    size_graph = decision_diagram.size();
     init_table();
     calculate_H_min();
     cleanup_arcs();
@@ -77,6 +68,7 @@ PricerSolverBdd::PricerSolverBdd(GPtrArray*  _jobs,
 PricerSolverBdd::PricerSolverBdd(const PricerSolverBdd& src,
                                  GPtrArray*             _ordered_jobs)
     : PricerSolverBase(src),
+      decision_diagram(src.decision_diagram),
       size_graph(src.size_graph),
       nb_removed_edges(src.nb_removed_edges),
       nb_removed_nodes(src.nb_removed_nodes),
@@ -85,37 +77,10 @@ PricerSolverBdd::PricerSolverBdd(const PricerSolverBdd& src,
       original_model(src.original_model),
       H_max(src.H_max),
       H_min(src.H_min) {
-    PricerConstruct ps(_ordered_jobs);
-    decision_diagram = std::make_unique<DdStructure<>>(ps);
     remove_layers_init();
-    decision_diagram->compressBdd();
-    size_graph = decision_diagram->size();
+    decision_diagram.compressBdd();
+    size_graph = decision_diagram.size();
     init_table();
-    calculate_H_min();
-    cleanup_arcs();
-    // check_infeasible_arcs();
-    bottum_up_filtering();
-    topdown_filtering();
-    construct_mipgraph();
-    init_coeff_constraints();
-}
-
-PricerSolverBdd::PricerSolverBdd(const PricerSolverBdd& src)
-    : PricerSolverBase(src),
-      decision_diagram(new DdStructure<>(*src.decision_diagram)),
-      size_graph(src.size_graph),
-      nb_removed_edges(src.nb_removed_edges),
-      nb_removed_nodes(src.nb_removed_nodes),
-      ordered_jobs(src.ordered_jobs),
-      mip_graph(src.mip_graph),
-      original_model(src.original_model),
-      H_max(src.H_max),
-      H_min(src.H_min) {
-    remove_layers_init();
-    decision_diagram->compressBdd();
-    size_graph = decision_diagram->size();
-    init_table();
-    calculate_H_min();
     cleanup_arcs();
     // check_infeasible_arcs();
     bottum_up_filtering();
@@ -125,20 +90,21 @@ PricerSolverBdd::PricerSolverBdd(const PricerSolverBdd& src)
 }
 
 void PricerSolverBdd::calculate_H_min() {
-    auto p_sum = 0.0;
-    auto duration = g_ptr_array_new();
+    auto  p_sum = 0.0;
+    auto* duration = g_ptr_array_new();
     for (auto j = 0; j < convex_constr_id; j++) {
-        auto job = (Job*)g_ptr_array_index(jobs, j);
+        auto* job = static_cast<Job*>(jobs[j]);
         g_ptr_array_add(duration, job);
         p_sum += job->processing_time;
     }
     g_ptr_array_sort(duration, g_compare_duration);
+    std::span aux_duration{duration->pdata, duration->len};
 
     auto m = 0;
     auto tmp = p_sum;
     auto i = convex_constr_id;
     do {
-        auto job = (Job*)g_ptr_array_index(duration, i - 1);
+        auto* job = static_cast<Job*>(aux_duration[i - 1]);
         tmp -= job->processing_time;
         m++;
         i--;
@@ -151,12 +117,12 @@ void PricerSolverBdd::calculate_H_min() {
 
 void PricerSolverBdd::construct_mipgraph() {
     mip_graph.clear();
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
     auto  vertex_nodeid_list(get(boost::vertex_name_t(), mip_graph));
     auto  edge_type_list(get(boost::edge_weight_t(), mip_graph));
 
-    for (auto i = decision_diagram->topLevel(); i >= 0; i--) {
-        for (auto j = 0u; j < table[i].size(); j++) {
+    for (auto i = decision_diagram.topLevel(); i >= 0; i--) {
+        for (auto j = 0U; j < table[i].size(); j++) {
             if (NodeId(i, j) != 0
                 // && (table[i][j].calc_yes || table[i][j].calc_no)
             ) {
@@ -168,21 +134,19 @@ void PricerSolverBdd::construct_mipgraph() {
 
     auto count = 0;
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
-            if (it.branch[0] != 0 && it.calc_no) {
+            if (it.branch[0] != 0 && it.calc[0]) {
                 auto& n0 = table.node(it.branch[0]);
                 auto  a = add_edge(it.key, n0.key, mip_graph);
                 put(edge_type_list, a.first, false);
-                it.low_edge_key = count;
                 put(boost::edge_index_t(), mip_graph, a.first, count++);
             }
 
-            if (it.branch[1] != 0 && it.calc_yes) {
+            if (it.branch[1] != 0 && it.calc[1]) {
                 auto& n1 = table.node(it.branch[1]);
                 auto  a = add_edge(it.key, n1.key, mip_graph);
                 put(edge_type_list, a.first, true);
-                it.high_edge_key = count;
                 put(boost::edge_index_t(), mip_graph, a.first, count++);
             }
         }
@@ -193,36 +157,38 @@ void PricerSolverBdd::construct_mipgraph() {
 }
 
 void PricerSolverBdd::init_coeff_constraints() {
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
+    original_model.clear_all_coeff();
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
+            it.add_coeff_list_clear();
             for (auto c = 0; c < reformulation_model.get_nb_constraints();
                  c++) {
                 if (c == convex_constr_id) {
                     continue;
                 }
 
-                auto            constr = reformulation_model.get_constraint(c);
+                auto*           constr = reformulation_model.get_constraint(c);
                 VariableKeyBase key_aux(it.get_nb_job(), it.get_weight());
                 auto            coeff = constr->get_var_coeff(&key_aux);
-                if (fabs(coeff) > 1e-10) {
+                if (fabs(coeff) > EPS_SOLVER) {
                     auto ptr_coeff{std::make_shared<BddCoeff>(
                         it.get_nb_job(), it.get_weight(), coeff, 0.0, c)};
                     original_model.add_coeff_list(c, ptr_coeff);
-                    it.add_coeff_list(ptr_coeff, 1);
+                    it.add_coeff_list(ptr_coeff, true);
                 }
             }
         }
     }
 
-    auto&           root = decision_diagram->root();
+    auto&           root = decision_diagram.root();
     auto&           root_node = table.node(root);
     VariableKeyBase key_aux(root_node.get_nb_job(), root_node.get_weight(),
                             true);
-    auto            constr = original_model.get_constraint(convex_constr_id);
+    auto*           constr = original_model.get_constraint(convex_constr_id);
     auto            coeff = constr->get_var_coeff(&key_aux);
-    if (fabs(coeff) > 1e-10) {
+    if (fabs(coeff) > EPS_SOLVER) {
         auto ptr_coeff_high{std::make_shared<BddCoeff>(
             root_node.get_nb_job(), root_node.get_weight(), coeff, true, true)};
         original_model.add_coeff_list(convex_constr_id, ptr_coeff_high);
@@ -243,20 +209,20 @@ void PricerSolverBdd::update_coeff_constraints() {
             reformulation_model.get_constraint_ptr(j));
     }
 
-    auto& table = *(decision_diagram->getDiagram());
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    auto& table = *(decision_diagram.getDiagram());
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             for (int c = 0; c < nb_new_constr; c++) {
-                auto     constr = original_model.get_constraint(nb_constr + c);
+                auto*    constr = original_model.get_constraint(nb_constr + c);
                 BddCoeff key_high{it.get_nb_job(), it.get_weight(), 0.0, 0.0,
                                   nb_constr + c};
                 auto     coeff_high = constr->get_var_coeff(&key_high);
-                if (fabs(coeff_high) > 1e-10) {
+                if (fabs(coeff_high) > EPS_SOLVER) {
                     auto ptr_coeff{std::make_shared<BddCoeff>(
                         it.get_nb_job(), it.get_weight(), coeff_high, 0.0,
                         nb_constr + c)};
                     original_model.add_coeff_list(c + nb_constr, ptr_coeff);
-                    it.add_coeff_list(ptr_coeff, 1);
+                    it.add_coeff_list(ptr_coeff, true);
                 }
 
                 BddCoeff key_low{it.get_nb_job(),
@@ -266,13 +232,13 @@ void PricerSolverBdd::update_coeff_constraints() {
                                  nb_constr + c,
                                  false};
                 auto     coeff_low = constr->get_var_coeff(&key_low);
-                if (fabs(coeff_low) > 1e-10) {
+                if (fabs(coeff_low) > EPS_SOLVER) {
                     std::shared_ptr<BddCoeff> ptr_coeff{
                         std::make_shared<BddCoeff>(it.get_nb_job(),
                                                    it.get_weight(), coeff_low,
                                                    0.0, nb_constr + c, false)};
                     original_model.add_coeff_list(c + nb_constr, ptr_coeff);
-                    it.add_coeff_list(ptr_coeff, 0);
+                    it.add_coeff_list(ptr_coeff, false);
                 }
             }
         }
@@ -280,22 +246,23 @@ void PricerSolverBdd::update_coeff_constraints() {
 }
 
 void PricerSolverBdd::init_table() {
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
     /** init table */
-    auto& root = table.node(decision_diagram->root());
+    auto& root = table.node(decision_diagram.root());
     root.init_node(0, true);
-    root.set_node_id_label(decision_diagram->root());
+    root.set_node_id_label(decision_diagram.root());
     root.all = boost::dynamic_bitset<>{convex_constr_id, 0};
+    std::span aux_ordered_jobs{ordered_jobs->pdata, ordered_jobs->len};
 
-    for (auto i = decision_diagram->topLevel(); i >= 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i >= 0; i--) {
         for (auto it = 0u; it < table[i].size(); it++) {
             if (i != 0) {
-                auto layer = ordered_jobs->len - i;
-                auto tmp_pair = static_cast<job_interval_pair*>(
-                    g_ptr_array_index(ordered_jobs, layer));
+                auto  layer = ordered_jobs->len - i;
+                auto* tmp_pair =
+                    static_cast<job_interval_pair*>(aux_ordered_jobs[layer]);
                 auto  node_id = NodeId(i, it);
                 auto& node = table.node(node_id);
-                auto  aux_job = tmp_pair->j;
+                auto* aux_job = tmp_pair->j;
                 auto  w = node.get_weight();
                 auto  p = aux_job->processing_time;
                 // node.set_node_id_label(node_id);
@@ -312,30 +279,29 @@ void PricerSolverBdd::init_table() {
                 node.cost[0] = 0.0;
                 node.cost[1] = value_Fj(w + p, aux_job);
 
-                n0.in_degree_0++;
+                n0.in_degree[0]++;
                 n0.in_edges[0].push_back(node.ptr_node_id);
-                n1.in_degree_1++;
+                n1.in_degree[1]++;
                 n1.in_edges[1].push_back(node.ptr_node_id);
 
-                auto iter = t_out.find(w);
-                if (iter == t_out.end()) {
-                    t_out[w] =
-                        std::vector<std::weak_ptr<NodeId>>{node.ptr_node_id};
-                } else {
-                    iter->second.push_back(node.ptr_node_id);
-                }
+                // auto iter = t_out.find(w);
+                // if (iter == t_out.end()) {
+                //     t_out[w] =
+                //         std::vector<std::weak_ptr<NodeId>>{node.ptr_node_id};
+                // } else {
+                //     iter->second.push_back(node.ptr_node_id);
+                // }
 
-                iter = t_in.find(w + p);
-                if (iter == t_in.end()) {
-                    t_in[w + p] =
-                        std::vector<std::weak_ptr<NodeId>>{node.ptr_node_id};
-                } else {
-                    iter->second.push_back(node.ptr_node_id);
-                }
+                // iter = t_in.find(w + p);
+                // if (iter == t_in.end()) {
+                //     t_in[w + p] =
+                //         std::vector<std::weak_ptr<NodeId>>{node.ptr_node_id};
+                // } else {
+                //     iter->second.push_back(node.ptr_node_id);
+                // }
 
             } else {
                 auto& node = table.node(NodeId(i, it));
-                // node.set_node_id_label(NodeId(i, it));
                 node.set_job(nullptr);
             }
         }
@@ -344,14 +310,15 @@ void PricerSolverBdd::init_table() {
 
 void PricerSolverBdd::insert_constraints_lp(NodeData* pd) {
     lp_interface_get_nb_rows(pd->RMP, &(pd->nb_rows));
-    int nb_new_constraints =
+    auto nb_new_constraints =
         reformulation_model.get_nb_constraints() - pd->nb_rows;
+    std::span aux_col_pool{pd->localColPool->pdata, pd->localColPool->len};
 
     fmt::print("nb rows initial {} {} {}\n", pd->nb_rows,
                reformulation_model.get_nb_constraints(), nb_new_constraints);
 
-    assert(nb_new_constraints <=
-           (pd->id_pseudo_schedules - pd->id_next_var_cuts));
+    assert((nb_new_constraints <=
+            (pd->id_pseudo_schedules - pd->id_next_var_cuts)));
     std::vector<int>    starts(nb_new_constraints + 1);
     std::vector<char>   sense(nb_new_constraints);
     std::vector<double> rhs(nb_new_constraints);
@@ -360,7 +327,7 @@ void PricerSolverBdd::insert_constraints_lp(NodeData* pd) {
 
     int pos = 0;
     for (int c = 0; c < nb_new_constraints; c++) {
-        auto constr = reformulation_model.get_constraint(pd->nb_rows + c);
+        auto* constr = reformulation_model.get_constraint(pd->nb_rows + c);
 
         sense[c] = constr->get_sense();
         starts[c] = pos;
@@ -376,21 +343,21 @@ void PricerSolverBdd::insert_constraints_lp(NodeData* pd) {
             }
         }
 
-        for (auto i = 0u; i < pd->localColPool->len; i++) {
-            auto aux_schedule_set =
-                (ScheduleSet*)g_ptr_array_index(pd->localColPool, i);
-            auto  jobs_list = aux_schedule_set->job_list;
-            auto& table = *(decision_diagram->getDiagram());
-            auto  tmp_nodeid(decision_diagram->root());
+        for (auto i = 0U; i < pd->localColPool->len; i++) {
+            auto* aux_schedule_set = static_cast<ScheduleSet*>(aux_col_pool[i]);
+            std::span jobs_list{aux_schedule_set->job_list->pdata,
+                                aux_schedule_set->job_list->len};
+            auto&     table = *(decision_diagram.getDiagram());
+            auto      tmp_nodeid(decision_diagram.root());
 
             auto coeff_val = 0.0;
-            auto counter = 0u;
+            auto counter = 0U;
             while (tmp_nodeid > 1) {
                 auto& tmp_node = table.node(tmp_nodeid);
                 Job*  tmp_j = nullptr;
 
-                if (counter < jobs_list->len) {
-                    tmp_j = (Job*)g_ptr_array_index(jobs_list, counter);
+                if (counter < jobs_list.size()) {
+                    tmp_j = static_cast<Job*>(jobs_list[counter]);
                 }
 
                 VariableKeyBase key(tmp_node.get_nb_job(),
@@ -406,9 +373,9 @@ void PricerSolverBdd::insert_constraints_lp(NodeData* pd) {
                 }
             }
 
-            assert(tmp_nodeid == 1);
+            assert((tmp_nodeid == 1));
 
-            if (fabs(coeff_val) > 1e-6) {
+            if (fabs(coeff_val) > EPS_SOLVER) {
                 column_ind.push_back(pd->id_pseudo_schedules + i);
                 coeff.push_back(coeff_val);
                 pos++;
@@ -422,16 +389,15 @@ void PricerSolverBdd::insert_constraints_lp(NodeData* pd) {
                          starts.data(), column_ind.data(), coeff.data(),
                          sense.data(), rhs.data(), nullptr);
     lp_interface_write(pd->RMP, "test_build.lp");
-    int test;
-    lp_interface_get_nb_rows(pd->RMP, &(test));
-    pd->nb_rows = test;
-
+    lp_interface_get_nb_rows(pd->RMP, &(pd->nb_rows));
     vector<double> new_values(nb_new_constraints, 0.0);
     vector<int>    new_values_int(nb_new_constraints, 0);
     g_array_append_vals(pd->pi, new_values.data(), new_values.size());
     g_array_append_vals(pd->slack, new_values.data(), new_values.size());
     g_array_append_vals(pd->rhs, new_values.data(), new_values.size());
-    lp_interface_get_rhs(pd->RMP, &g_array_index(pd->rhs, double, 0));
+    lp_interface_get_rhs(
+        pd->RMP, static_cast<double*>(static_cast<void*>(pd->rhs->data)));
+    // lp_interface_get_rhs(pd->RMP, &g_array_index(pd->rhs, double, 0));
     g_array_append_vals(pd->lhs_coeff, new_values.data(), new_values.size());
     g_array_append_vals(pd->id_row, new_values_int.data(),
                         new_values_int.size());
@@ -443,17 +409,20 @@ double PricerSolverBdd::compute_reduced_cost(const OptimalSolution<>& sol,
                                              double*                  pi,
                                              double*                  lhs) {
     double result = sol.cost;
-    auto&  table = *decision_diagram->getDiagram();
-    auto   tmp_nodeid(decision_diagram->root());
-    auto   counter = 0u;
+    auto&  table = *decision_diagram.getDiagram();
+    auto   tmp_nodeid(decision_diagram.root());
+    auto   counter = 0U;
 
-    std::fill(lhs, lhs + reformulation_model.get_nb_constraints(), 0.0);
+    std::span aux_lhs{lhs, reformulation_model.get_nb_constraints()};
+    std::span aux_pi{pi, reformulation_model.get_nb_constraints()};
+    std::span aux_jobs{sol.jobs->pdata, sol.jobs->len};
+    std::fill(aux_lhs.begin(), aux_lhs.end(), 0.0);
     while (tmp_nodeid > 1) {
         auto& tmp_node = table.node(tmp_nodeid);
         Job*  tmp_j = nullptr;
 
         if (counter < sol.jobs->len) {
-            tmp_j = (Job*)g_ptr_array_index(sol.jobs, counter);
+            tmp_j = static_cast<Job*>(aux_jobs[counter]);
         }
 
         VariableKeyBase key(tmp_node.get_nb_job(), tmp_node.get_weight(),
@@ -461,13 +430,13 @@ double PricerSolverBdd::compute_reduced_cost(const OptimalSolution<>& sol,
         if (key.get_high()) {
             tmp_nodeid = tmp_node.branch[1];
             counter++;
-            auto dual = pi[key.get_j()];
-            auto constr = reformulation_model.get_constraint(key.get_j());
-            auto coeff = constr->get_var_coeff(&key);
+            auto* constr = reformulation_model.get_constraint(key.get_j());
+            auto  dual = aux_pi[key.get_j()];
+            auto  coeff = constr->get_var_coeff(&key);
 
-            if (fabs(coeff) > 1e-10) {
+            if (fabs(coeff) > EPS_SOLVER) {
                 result -= coeff * dual;
-                lhs[key.get_j()] += coeff;
+                aux_lhs[key.get_j()] += coeff;
             }
         } else {
             tmp_nodeid = tmp_node.branch[0];
@@ -475,23 +444,23 @@ double PricerSolverBdd::compute_reduced_cost(const OptimalSolution<>& sol,
 
         for (int c = convex_constr_id + 1;
              c < reformulation_model.get_nb_constraints(); c++) {
-            auto dual = pi[c];
-            auto constr = reformulation_model.get_constraint(c);
-            auto coeff = constr->get_var_coeff(&key);
+            auto* constr = reformulation_model.get_constraint(c);
+            auto  dual = aux_pi[c];
+            auto  coeff = constr->get_var_coeff(&key);
 
-            if (fabs(coeff) > 1e-10) {
+            if (fabs(coeff) > EPS_SOLVER) {
                 result -= coeff * dual;
-                lhs[c] += coeff;
+                aux_lhs[c] += coeff;
             }
         }
     }
 
-    auto dual = pi[convex_constr_id];
-    auto constr = reformulation_model.get_constraint(convex_constr_id);
+    auto* constr = reformulation_model.get_constraint(convex_constr_id);
+    auto  dual = aux_pi[convex_constr_id];
     VariableKeyBase k(0, 0, true);
     auto            coeff = constr->get_var_coeff(&k);
     result -= coeff * dual;
-    lhs[convex_constr_id] += coeff;
+    aux_lhs[convex_constr_id] += coeff;
 
     return result;
 }
@@ -499,16 +468,18 @@ double PricerSolverBdd::compute_reduced_cost(const OptimalSolution<>& sol,
 double PricerSolverBdd::compute_subgradient(const OptimalSolution<>& sol,
                                             double* sub_gradient) {
     double result = sol.cost;
-    auto&  table = *decision_diagram->getDiagram();
-    auto   tmp_nodeid(decision_diagram->root());
-    auto   counter = 0u;
+    auto&  table = *decision_diagram.getDiagram();
+    auto   tmp_nodeid(decision_diagram.root());
+    auto   counter = 0U;
     auto   nb_constraints = reformulation_model.get_nb_constraints();
     auto   convex_rhs =
         -reformulation_model.get_constraint(convex_constr_id)->get_rhs();
+    std::span aux_subgradient{sub_gradient, nb_constraints};
+    std::span aux_jobs{sol.jobs->pdata, sol.jobs->len};
 
     for (size_t i = 0; i < nb_constraints; i++) {
-        auto constr = reformulation_model.get_constraint(i);
-        sub_gradient[i] = constr->get_rhs();
+        auto* constr = reformulation_model.get_constraint(i);
+        aux_subgradient[i] = constr->get_rhs();
     }
 
     while (tmp_nodeid > 1) {
@@ -516,7 +487,7 @@ double PricerSolverBdd::compute_subgradient(const OptimalSolution<>& sol,
         Job*  tmp_j = nullptr;
 
         if (counter < sol.jobs->len) {
-            tmp_j = (Job*)g_ptr_array_index(sol.jobs, counter);
+            tmp_j = static_cast<Job*>(aux_jobs[counter]);
         }
 
         VariableKeyBase key(tmp_node.get_nb_job(), tmp_node.get_weight(),
@@ -524,11 +495,11 @@ double PricerSolverBdd::compute_subgradient(const OptimalSolution<>& sol,
         if (key.get_high()) {
             tmp_nodeid = tmp_node.branch[1];
             counter++;
-            auto constr = reformulation_model.get_constraint(key.get_j());
-            auto coeff = constr->get_var_coeff(&key);
+            auto* constr = reformulation_model.get_constraint(key.get_j());
+            auto  coeff = constr->get_var_coeff(&key);
 
-            if (fabs(coeff) > 1e-10) {
-                sub_gradient[key.get_j()] -= coeff * convex_rhs;
+            if (fabs(coeff) > EPS_SOLVER) {
+                aux_subgradient[key.get_j()] -= coeff * convex_rhs;
             }
         } else {
             tmp_nodeid = tmp_node.branch[0];
@@ -537,17 +508,17 @@ double PricerSolverBdd::compute_subgradient(const OptimalSolution<>& sol,
         for (int c = convex_constr_id + 1;
              c < reformulation_model.get_nb_constraints(); c++) {
             // auto dual = pi[c];
-            auto constr = reformulation_model.get_constraint(c);
-            auto coeff = constr->get_var_coeff(&key);
+            auto* constr = reformulation_model.get_constraint(c);
+            auto  coeff = constr->get_var_coeff(&key);
 
-            if (fabs(coeff) > 1e-10) {
-                sub_gradient[c] -= coeff * convex_rhs;
+            if (fabs(coeff) > EPS_SOLVER) {
+                aux_subgradient[c] -= coeff * convex_rhs;
             }
         }
     }
 
-    sub_gradient[convex_constr_id] += convex_rhs;
-    assert(sub_gradient[convex_constr_id] == 0.0);
+    aux_subgradient[convex_constr_id] += convex_rhs;
+    assert(aux_subgradient[convex_constr_id] == 0.0);
     // sub_gradient[nb_jobs] = 0.0;
     assert(tmp_nodeid == 1);
 
@@ -559,26 +530,29 @@ double PricerSolverBdd::compute_lagrange(const OptimalSolution<>& sol,
     double result = sol.cost;
     auto   dual_bound = 0.0;
 
-    auto& table = *decision_diagram->getDiagram();
-    auto  tmp_nodeid(decision_diagram->root());
+    auto&     table = *decision_diagram.getDiagram();
+    auto      tmp_nodeid(decision_diagram.root());
+    auto      nb_constraints = reformulation_model.get_nb_constraints();
+    std::span aux_pi{pi, nb_constraints};
+    std::span aux_jobs{sol.jobs->pdata, sol.jobs->len};
 
-    auto counter = 0u;
+    auto counter = 0U;
     while (tmp_nodeid > 1) {
         auto& tmp_node = table.node(tmp_nodeid);
         Job*  tmp_j = nullptr;
 
         if (counter < sol.jobs->len) {
-            tmp_j = (Job*)g_ptr_array_index(sol.jobs, counter);
+            tmp_j = static_cast<Job*>(aux_jobs[counter]);
         }
 
         VariableKeyBase key(tmp_node.get_nb_job(), tmp_node.get_weight(),
                             tmp_j == tmp_node.get_job());
         if (key.get_high()) {
-            auto dual = pi[key.get_j()];
-            auto constr = reformulation_model.get_constraint(key.get_j());
-            auto coeff = constr->get_var_coeff(&key);
+            auto* constr = reformulation_model.get_constraint(key.get_j());
+            auto  dual = aux_pi[key.get_j()];
+            auto  coeff = constr->get_var_coeff(&key);
 
-            if (fabs(coeff) > 1e-10) {
+            if (fabs(coeff) > EPS_SOLVER) {
                 result -= coeff * dual;
             }
 
@@ -593,11 +567,11 @@ double PricerSolverBdd::compute_lagrange(const OptimalSolution<>& sol,
             if (c == convex_constr_id) {
                 continue;
             }
-            auto dual = pi[c];
-            auto constr = reformulation_model.get_constraint(c);
-            auto coeff = constr->get_var_coeff(&key);
+            auto* constr = reformulation_model.get_constraint(c);
+            auto  dual = aux_pi[c];
+            auto  coeff = constr->get_var_coeff(&key);
 
-            if (fabs(coeff) > 1e-10) {
+            if (fabs(coeff) > EPS_SOLVER) {
                 result -= coeff * dual;
             }
         }
@@ -609,9 +583,9 @@ double PricerSolverBdd::compute_lagrange(const OptimalSolution<>& sol,
         if (c == convex_constr_id) {
             continue;
         }
-        auto dual = pi[c];
-        auto constr = reformulation_model.get_constraint(c);
-        auto rhs = constr->get_rhs();
+        auto* constr = reformulation_model.get_constraint(c);
+        auto  dual = aux_pi[c];
+        auto  rhs = constr->get_rhs();
 
         dual_bound += rhs * dual;
     }
@@ -628,10 +602,10 @@ void PricerSolverBdd::remove_layers_init() {
     auto  first_del = -1;
     auto  last_del = -1;
     auto  it = 0;
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
 
     /** remove the unnecessary layers of the bdd */
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         if (std::any_of(table[i].begin(), table[i].end(),
                         [](NodeBdd<>& n) { return n.branch[1] != 0; })) {
             if (first_del != -1) {
@@ -661,7 +635,7 @@ void PricerSolverBdd::remove_layers_init() {
 
     if (dbg_lvl() > 0) {
         fmt::print("{0: <{2}}{1}\n", "The new number of layers",
-                   ordered_jobs->len, 60);
+                   ordered_jobs->len, ALIGN);
     }
 }
 
@@ -669,14 +643,14 @@ void PricerSolverBdd::remove_layers() {
     auto  first_del = -1;
     auto  last_del = -1;
     auto  it = 0;
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
 
     /** remove the unnecessary layers of the bdd */
-    for (int i = decision_diagram->topLevel(); i > 0; i--) {
+    for (int i = decision_diagram.topLevel(); i > 0; i--) {
         auto remove = true;
 
         for (auto& iter : table[i]) {
-            if (iter.calc_yes) {
+            if (iter.calc[1]) {
                 remove = false;
             } else {
                 auto& cur_node_1 = iter.branch[1];
@@ -712,40 +686,41 @@ void PricerSolverBdd::remove_layers() {
 
     if (dbg_lvl() > 0) {
         fmt::print("{0: <{2}}{1}\n", "The new number of layers",
-                   ordered_jobs->len, 60);
+                   ordered_jobs->len, ALIGN);
     }
 }
 
 void PricerSolverBdd::remove_edges() {
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
 
     /** remove the unnecessary nodes of the bdd */
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& iter : table[i]) {
-            if (!iter.calc_yes) {
-                NodeId& cur_node_1 = iter.branch[1];
+            if (!iter.calc[1]) {
+                auto& cur_node_1 = iter.branch[1];
                 iter.ptr_node_id.reset();
                 cur_node_1 = 0;
             }
 
-            if (!iter.calc_no) {
-                NodeId& cur_node_0 = iter.branch[0];
+            if (!iter.calc[0]) {
+                auto& cur_node_0 = iter.branch[0];
                 cur_node_0 = 0;
             }
         }
     }
 
-    decision_diagram->compressBdd();
+    decision_diagram.compressBdd();
     nb_removed_nodes -= size_graph;
-    size_graph = decision_diagram->size();
+    size_graph = decision_diagram.size();
 }
 
 void PricerSolverBdd::print_representation_file() {
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
     auto  vertex_nodeid_list(get(boost::vertex_name_t(), mip_graph));
     auto  edge_type_list(get(boost::edge_weight_t(), mip_graph));
     auto  edge_index_list(get(boost::edge_index_t(), mip_graph));
-    auto  index_edge{std::make_unique<std::vector<int>[]>(convex_constr_id)};
+    auto  index_edge{
+        std::vector<std::vector<int>>(convex_constr_id, std::vector<int>())};
 
     auto outfile_file_mip_str =
         problem_name + "_" + std::to_string(convex_rhs) + ".txt";
@@ -812,7 +787,7 @@ void PricerSolverBdd::add_inequality(std::vector<int> v1, std::vector<int> v2) {
             expr1 -= edge_var_list[*it.first].x;
         }
     }
-    model->addConstr(expr1, GRB_EQUAL, 0);
+    model.addConstr(expr1, GRB_EQUAL, 0);
 }
 
 void PricerSolverBdd::add_inequality(std::vector<int> v1) {
@@ -825,17 +800,16 @@ void PricerSolverBdd::add_inequality(std::vector<int> v1) {
             expr1 += edge_var_list[*it.first].x;
         }
     }
-    model->addConstr(expr1, GRB_EQUAL, convex_rhs);
+    model.addConstr(expr1, GRB_EQUAL, convex_rhs);
 }
 void PricerSolverBdd::build_mip() {
     try {
         fmt::print("Building Mip model for the extended formulation:\n");
-        auto& table = *(decision_diagram->getDiagram());
+        auto& table = *(decision_diagram.getDiagram());
         auto  vertex_index_list(get(boost::vertex_index_t(), mip_graph));
         auto  vertex_nodeid_list(get(boost::vertex_name_t(), mip_graph));
         auto  edge_type_list{get(boost::edge_weight_t(), mip_graph)};
         auto  edge_var_list{get(boost::edge_weight2_t(), mip_graph)};
-        auto  edge_index_list{get(boost::edge_index_t(), mip_graph)};
 
         /** Constructing variables */
         for (auto it = edges(mip_graph); it.first != it.second; it.first++) {
@@ -845,23 +819,24 @@ void PricerSolverBdd::build_mip() {
                 auto   C = n.get_weight() + n.get_job()->processing_time;
                 double cost = value_Fj(C, n.get_job());
                 edge_var_list[*it.first].x =
-                    model->addVar(0.0, 1.0, cost, GRB_BINARY);
+                    model.addVar(0.0, 1.0, cost, GRB_BINARY);
             } else {
-                edge_var_list[*it.first].x = model->addVar(
+                edge_var_list[*it.first].x = model.addVar(
                     0.0, static_cast<double>(convex_rhs), 0.0, GRB_CONTINUOUS);
             }
         }
 
-        model->update();
+        model.update();
         /** Assignment constraints */
-        auto assignment{std::make_unique<GRBLinExpr[]>(convex_constr_id)};
-        auto sense{std::make_unique<char[]>(convex_constr_id)};
-        auto rhs{std::make_unique<double[]>(convex_constr_id)};
+        auto assignment{
+            std::vector<GRBLinExpr>(convex_constr_id, GRBLinExpr())};
+        auto sense{std::vector<char>(convex_constr_id, GRB_EQUAL)};
+        auto rhs{std::vector<double>(convex_constr_id, 1.0)};
 
-        for (unsigned i = 0; i < jobs->len; ++i) {
-            sense[i] = GRB_EQUAL;
-            rhs[i] = 1.0;
-        }
+        // for (unsigned i = 0; i < jobs->len; ++i) {
+        //     sense[i] = GRB_EQUAL;
+        //     rhs[i] = 1.0;
+        // }
 
         for (auto it = edges(mip_graph); it.first != it.second; it.first++) {
             auto high = edge_type_list[*it.first];
@@ -873,16 +848,16 @@ void PricerSolverBdd::build_mip() {
             }
         }
 
-        std::unique_ptr<GRBConstr[]> assignment_constrs(
-            model->addConstrs(assignment.get(), sense.get(), rhs.get(), nullptr,
-                              convex_constr_id));
-        model->update();
+        std::unique_ptr<GRBConstr> assignment_constrs(
+            model.addConstrs(assignment.data(), sense.data(), rhs.data(),
+                             nullptr, convex_constr_id));
+        model.update();
         /** Flow constraints */
         auto num_vertices = boost::num_vertices(mip_graph);
         auto flow_conservation_constr{
-            std::make_unique<GRBLinExpr[]>(num_vertices)};
-        auto sense_flow{std::make_unique<char[]>(num_vertices)};
-        auto rhs_flow(std::make_unique<double[]>(num_vertices));
+            std::vector<GRBLinExpr>(num_vertices, GRBLinExpr())};
+        auto sense_flow{std::vector<char>(num_vertices, GRB_EQUAL)};
+        auto rhs_flow(std::vector<double>(num_vertices, 0));
 
         for (auto it = vertices(mip_graph); it.first != it.second; ++it.first) {
             const auto node_id = vertex_nodeid_list[*it.first];
@@ -904,7 +879,7 @@ void PricerSolverBdd::build_mip() {
                     edge_var_list[*in_edges_it.first].x;
             }
 
-            if (node_id == decision_diagram->root()) {
+            if (node_id == decision_diagram.root()) {
                 rhs_flow[vertex_key] = static_cast<double>(-convex_rhs);
             } else if (node_id == 1) {
                 rhs_flow[vertex_key] = static_cast<double>(convex_rhs);
@@ -913,10 +888,10 @@ void PricerSolverBdd::build_mip() {
             }
         }
 
-        std::unique_ptr<GRBConstr[]> flow_constrs(
-            model->addConstrs(flow_conservation_constr.get(), sense_flow.get(),
-                              rhs_flow.get(), nullptr, num_vertices));
-        model->update();
+        std::unique_ptr<GRBConstr> flow_constrs(
+            model.addConstrs(flow_conservation_constr.data(), sense_flow.data(),
+                             rhs_flow.data(), nullptr, num_vertices));
+        model.update();
         // for (auto it = edges(mip_graph); it.first != it.second; it.first++) {
         //     // edge_var_list[*it.first].x.set(GRB_DoubleAttr_PStart,
         //     // lp_x[edge_index_list[*it.first]]);
@@ -924,26 +899,27 @@ void PricerSolverBdd::build_mip() {
         //         GRB_DoubleAttr_Start,
         //         solution_x[edge_index_list[*it.first]]);
         // }
-        model->write("original_" + problem_name + ".lp");
-        auto presolve = model->presolve();
+        model.write("original_" + problem_name + ".lp");
+        auto presolve = model.presolve();
         presolve.write("presolve_" + problem_name + ".lp");
-        model->optimize();
+        model.optimize();
 
-        if (model->get(GRB_IntAttr_Status) == GRB_OPTIMAL) {
-            for (auto it = edges(mip_graph); it.first != it.second;
-                 it.first++) {
-                int index = edge_index_list[*it.first];
-            }
+        // if (model.get(GRB_IntAttr_Status) == GRB_OPTIMAL) {
+        //     for (auto it = edges(mip_graph); it.first != it.second;
+        //          it.first++) {
+        //         int index = edge_index_list[*it.first];
+        //     }
 
-            // ColorWriterEdgeX  edge_writer(mip_graph, solution_x.get());
-            // ColorWriterVertex vertex_writer(mip_graph, table);
-            // string            file_name = "lp_solution_" + problem_name + "_"
-            // +
-            //                    std::to_string(num_machines) + ".gv";
-            // std::ofstream outf(file_name);
-            // boost::write_graphviz(outf, mip_graph, vertex_writer,
-            // edge_writer); outf.close();
-        }
+        //     // ColorWriterEdgeX  edge_writer(mip_graph, solution_x.get());
+        //     // ColorWriterVertex vertex_writer(mip_graph, table);
+        //     // string            file_name = "lp_solution_" + problem_name +
+        //     "_"
+        //     // +
+        //     //                    std::to_string(num_machines) + ".gv";
+        //     // std::ofstream outf(file_name);
+        //     // boost::write_graphviz(outf, mip_graph, vertex_writer,
+        //     // edge_writer); outf.close();
+        // }
 
         ColorWriterEdgeIndex edge_writer_index(mip_graph);
         ColorWriterVertex    vertex_writer(mip_graph, table);
@@ -1026,7 +1002,7 @@ void PricerSolverBdd::reduce_cost_fixing(double* pi, int UB, double LB) {
 }
 
 void PricerSolverBdd::cleanup_arcs() {
-    NodeTableEntity<>& table = *(decision_diagram->getDiagram());
+    NodeTableEntity<>& table = *(decision_diagram.getDiagram());
 
     table.node(0).backward_distance[0] = INT_MIN;
     table.node(0).backward_distance[1] = INT_MIN;
@@ -1035,10 +1011,9 @@ void PricerSolverBdd::cleanup_arcs() {
     auto removed_edges = false;
     auto nb_edges_removed_tmp = 0;
 
-    for (auto i = 1; i <= decision_diagram->topLevel(); i++) {
+    for (auto i = 1; i <= decision_diagram.topLevel(); i++) {
         for (auto& it : table[i]) {
-            it.calc_no = true;
-            it.calc_yes = true;
+            it.calc = {true, true};
             NodeBdd<>& cur_node_0 = table.node(it.branch[0]);
             NodeBdd<>& cur_node_1 = table.node(it.branch[1]);
 
@@ -1062,18 +1037,18 @@ void PricerSolverBdd::cleanup_arcs() {
         }
     }
     /** remove the unnecessary nodes of the bdd */
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& iter : table[i]) {
             if (iter.get_weight() + iter.backward_distance[0] < H_min &&
                 iter.branch[0] != 0) {
-                iter.calc_no = false;
+                iter.calc[0] = false;
                 removed_edges = true;
                 nb_edges_removed_tmp++;
                 nb_removed_edges++;
             }
 
             if (iter.get_weight() + iter.backward_distance[1] < H_min) {
-                iter.calc_yes = false;
+                iter.calc[1] = false;
                 removed_edges = true;
                 nb_edges_removed_tmp++;
                 nb_removed_edges++;
@@ -1084,9 +1059,9 @@ void PricerSolverBdd::cleanup_arcs() {
     if (removed_edges) {
         if (dbg_lvl() > 0) {
             fmt::print("{0: <{2}}{1}\n", "Number of edges removed by clean up",
-                       nb_edges_removed_tmp, 60);
+                       nb_edges_removed_tmp, ALIGN);
             fmt::print("{0: <{2}}{1}\n", "Total number of edges removed",
-                       get_nb_removed_edges(), 60);
+                       get_nb_removed_edges(), ALIGN);
         }
         remove_layers();
         remove_edges();
@@ -1096,18 +1071,18 @@ void PricerSolverBdd::cleanup_arcs() {
 void PricerSolverBdd::topdown_filtering() {
     auto  removed_edges = false;
     auto  nb_edges_removed_tmp = 0;
-    auto& table = *(decision_diagram->getDiagram());
-    auto& root = table.node(decision_diagram->root());
+    auto& table = *(decision_diagram.getDiagram());
+    auto& root = table.node(decision_diagram.root());
     root.init_node(0, true);
-    for (auto i = decision_diagram->topLevel(); i >= 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i >= 0; i--) {
         for (auto& it : table[i]) {
             it.visited = false;
             it.all = boost::dynamic_bitset<>{convex_constr_id, 0};
-            it.calc_yes = true;
+            it.calc[1] = true;
         }
     }
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             auto& n0 = table.node(it.branch[0]);
 
@@ -1124,24 +1099,24 @@ void PricerSolverBdd::topdown_filtering() {
             if (n1.visited) {
                 if (n1.all[it.get_nb_job()]) {
                     n1.all &= it.all;
-                    n1.all[it.get_nb_job()] = 1;
+                    n1.all[it.get_nb_job()] = true;
                 } else {
                     n1.all &= it.all;
                 }
             } else {
                 n1.all = boost::dynamic_bitset<>{convex_constr_id, 0};
                 n1.all |= it.all;
-                n1.all[it.get_nb_job()] = 1;
+                n1.all[it.get_nb_job()] = true;
                 n1.visited = true;
             }
         }
     }
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             if (it.all[it.get_nb_job()]) {
                 removed_edges = true;
-                it.calc_yes = false;
+                it.calc[1] = false;
                 nb_removed_edges++;
                 nb_edges_removed_tmp++;
             }
@@ -1166,12 +1141,12 @@ void PricerSolverBdd::topdown_filtering() {
 void PricerSolverBdd::bottum_up_filtering() {
     auto  removed_edges = false;
     auto  nb_edges_removed_tmp = 0;
-    auto& table = *(decision_diagram->getDiagram());
-    for (int i = decision_diagram->topLevel(); i > 0; i--) {
+    auto& table = *(decision_diagram.getDiagram());
+    for (int i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             it.visited = false;
             it.all = boost::dynamic_bitset<>{convex_constr_id, 0};
-            it.calc_yes = true;
+            it.calc[1] = true;
         }
     }
 
@@ -1179,19 +1154,19 @@ void PricerSolverBdd::bottum_up_filtering() {
     table.node(1).all = boost::dynamic_bitset<>{convex_constr_id, 0};
     table.node(0).all.flip();
 
-    for (auto i = 1; i <= decision_diagram->topLevel(); i++) {
+    for (auto i = 1; i <= decision_diagram.topLevel(); i++) {
         for (auto& it : table[i]) {
-            it.all[it.get_nb_job()] = 1;
+            it.all[it.get_nb_job()] = true;
             it.all |= table.node(it.branch[1]).all;
             it.all &= table.node(it.branch[0]).all;
         }
     }
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             if (table.node(it.branch[1]).all[it.get_nb_job()]) {
                 removed_edges = true;
-                it.calc_yes = false;
+                it.calc[1] = false;
                 nb_removed_edges++;
                 nb_edges_removed_tmp++;
             }
@@ -1217,40 +1192,38 @@ void PricerSolverBdd::check_infeasible_arcs() {
     /** init table */
     auto  removed_edges = false;
     auto  nb_edges_removed_tmp = 0;
-    auto& table = *(decision_diagram->getDiagram());
-    for (auto i = decision_diagram->topLevel(); i >= 0; i--) {
+    auto& table = *(decision_diagram.getDiagram());
+    for (auto i = decision_diagram.topLevel(); i >= 0; i--) {
         for (auto& it : table[i]) {
             it.visited = false;
             it.all = boost::dynamic_bitset<>{convex_constr_id, 0};
-            it.calc_yes = true;
-            it.calc_no = true;
+            it.calc = {true, true};
         }
     }
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             auto& n0 = table.node(it.branch[0]);
             n0.all |= it.all;
             auto& n1 = table.node(it.branch[1]);
-            n1.all[it.get_nb_job()] = 1;
+            n1.all[it.get_nb_job()] = true;
         }
     }
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             if (!it.all.empty() &&
                 it.all.find_first() != boost::dynamic_bitset<>::npos) {
                 auto index = it.all.find_first();
 
                 auto max = value_diff_Fij(it.get_weight(), it.get_job(),
-                                          (Job*)g_ptr_array_index(jobs, index));
+                                          static_cast<Job*>(jobs[index]));
                 // bool index_bool = (index > (size_t)it.get_nb_job());
                 while (index != boost::dynamic_bitset<>::npos && max < 0) {
                     index = it.all.find_next(index);
                     if (index != boost::dynamic_bitset<>::npos) {
-                        int a = value_diff_Fij(
-                            it.get_weight(), it.get_job(),
-                            (Job*)g_ptr_array_index(jobs, index));
+                        int a = value_diff_Fij(it.get_weight(), it.get_job(),
+                                               static_cast<Job*>(jobs[index]));
                         if (a > max) {
                             max = a;
                         }
@@ -1259,7 +1232,7 @@ void PricerSolverBdd::check_infeasible_arcs() {
 
                 if (max < 0) {
                     removed_edges = true;
-                    it.calc_yes = false;
+                    it.calc[1] = false;
                     nb_removed_edges++;
                     nb_edges_removed_tmp++;
                 }
@@ -1283,25 +1256,24 @@ void PricerSolverBdd::equivalent_paths_filtering() {
     auto  removed_edges = false;
     auto  nb_edges_removed_tmp = 0;
     auto  edge_type_list(get(boost::edge_weight_t(), mip_graph));
-    auto& table = *(decision_diagram->getDiagram());
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    auto& table = *(decision_diagram.getDiagram());
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             it.visited = false;
             it.all = boost::dynamic_bitset<>{convex_constr_id, 0};
-            it.calc_yes = true;
-            it.calc_no = true;
+            it.calc = {true, true};
             auto& n0 = table.node(it.branch[0]);
-            n0.in_degree_0++;
+            n0.in_degree[0]++;
             auto& n1 = table.node(it.branch[1]);
-            n1.in_degree_1++;
+            n1.in_degree[1]++;
         }
     }
 
     std::vector<int> vertices;
 
-    for (int i = decision_diagram->topLevel(); i > 0; i--) {
+    for (int i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
-            if (it.in_degree_1 + it.in_degree_0 >= 2) {
+            if (it.in_degree[0] + it.in_degree[0] >= 2) {
                 vertices.push_back(it.key);
             }
         }
@@ -1313,15 +1285,11 @@ void PricerSolverBdd::equivalent_paths_filtering() {
         list<NodeId> queue;
 
         queue.push_back(start_v);
-        std::unique_ptr<bool[]> visited(new bool[num_vertices]());
-        std::unique_ptr<bool[]> edge_visited(new bool[num_vertices]());
-        std::unique_ptr<boost::dynamic_bitset<>[]> all(
-            new dynamic_bitset<>[num_vertices]);
-        std::unique_ptr<int[]> C(new int[num_vertices]);
-        for (auto i = 0u; i < num_vertices; i++) {
-            all[i] = dynamic_bitset<>(convex_constr_id, 0);
-            C[i] = 0;
-        }
+        std::vector<bool>                    visited(num_vertices, false);
+        std::vector<bool>                    edge_visited(num_vertices, false);
+        std::vector<boost::dynamic_bitset<>> all(
+            num_vertices, boost::dynamic_bitset(convex_constr_id, 0));
+        std::vector<int> C(num_vertices, 0);
 
         auto& tmp_n = table.node(start_v);
         visited[tmp_n.key];
@@ -1344,7 +1312,7 @@ void PricerSolverBdd::equivalent_paths_filtering() {
                     if (high) {
                         auto& tmp_node = table.node(n.branch[1]);
                         all[n.key] |= all[tmp_node.key];
-                        all[n.key][n.get_nb_job()] = 1;
+                        all[n.key][n.get_nb_job()] = true;
                         C[n.key] = C[tmp_node.key] +
                                    value_Fj(tmp_node.get_weight(), n.get_job());
                         edge_visited[n.key] = true;
@@ -1354,12 +1322,12 @@ void PricerSolverBdd::equivalent_paths_filtering() {
                         C[n.key] = C[tmp_node.key];
                     }
                 } else {
-                    dynamic_bitset<> tmp;
-                    int              tmp_C;
+                    boost::dynamic_bitset<> tmp;
+                    int                     tmp_C{};
                     if (high) {
                         auto& tmp_node = table.node(n.branch[1]);
                         tmp = all[tmp_node.key];
-                        tmp[n.get_nb_job()] = 1;
+                        tmp[n.get_nb_job()] = true;
                         tmp_C = C[tmp_node.key] +
                                 value_Fj(tmp_node.get_weight(), n.get_job());
                     } else {
@@ -1369,7 +1337,7 @@ void PricerSolverBdd::equivalent_paths_filtering() {
                     }
 
                     if (all[n.key] == tmp) {
-                        NodeId cur;
+                        NodeId cur{};
                         NodeId prev = adjVertex;
                         if (high) {
                             if (tmp_C > C[n.key]) {
@@ -1387,7 +1355,7 @@ void PricerSolverBdd::equivalent_paths_filtering() {
 
                         while (cur != start_v) {
                             auto& node = table.node(cur);
-                            if (node.in_degree_1 + node.in_degree_0 > 1) {
+                            if (node.in_degree[0] + node.in_degree[1] > 1) {
                                 break;
                             }
                             prev = cur;
@@ -1402,11 +1370,11 @@ void PricerSolverBdd::equivalent_paths_filtering() {
 
                         auto& node_delete = table.node(prev);
                         if (edge_visited[node_delete.key] && cur == start_v) {
-                            node_delete.calc_yes = false;
+                            node_delete.calc[1] = false;
                             removed_edges = true;
                             nb_edges_removed_tmp++;
                         } else if (cur == start_v) {
-                            node_delete.calc_no = false;
+                            node_delete.calc[0] = false;
                             removed_edges = true;
                             nb_edges_removed_tmp++;
                         }
@@ -1438,14 +1406,14 @@ void PricerSolverBdd::equivalent_paths_filtering() {
 void PricerSolverBdd::add_constraint(Job* job, GPtrArray* list, int order) {
     scheduling         constr(job, list, order);
     std::ofstream      outf("min1.gv");
-    NodeTableEntity<>& table = *(decision_diagram->getDiagram());
+    NodeTableEntity<>& table = *(decision_diagram.getDiagram());
     ColorWriterVertex  vertex_writer(mip_graph, table);
     boost::write_graphviz(outf, mip_graph, vertex_writer);
-    decision_diagram->zddSubset(constr);
+    decision_diagram.zddSubset(constr);
     outf.close();
-    decision_diagram->compressBdd();
+    decision_diagram.compressBdd();
     construct_mipgraph();
-    auto&             table1 = *(decision_diagram->getDiagram());
+    auto&             table1 = *(decision_diagram.getDiagram());
     ColorWriterVertex vertex_writer1(mip_graph, table1);
     outf = std::ofstream("min2.gv");
     boost::write_graphviz(outf, mip_graph, vertex_writer1);
@@ -1455,39 +1423,44 @@ void PricerSolverBdd::add_constraint(Job* job, GPtrArray* list, int order) {
 void PricerSolverBdd::construct_lp_sol_from_rmp(const double*    columns,
                                                 const GPtrArray* schedule_sets,
                                                 int              num_columns) {
-    auto& table = *(decision_diagram->getDiagram());
-    for (auto i = decision_diagram->topLevel(); i >= 0; i--) {
+    auto& table = *(decision_diagram.getDiagram());
+    for (auto i = decision_diagram.topLevel(); i >= 0; i--) {
         for (auto& it : table[i]) {
             it.reset_lp_x();
         }
     }
+    auto                    nb_columns = static_cast<size_t>(num_columns);
+    std::span<const double> aux_cols{columns, nb_columns};
+    std::span aux_schedule_sets{schedule_sets->pdata, schedule_sets->len};
+    assert(nb_columns == schedule_sets->len);
 
     set_is_integer_solution(true);
     for (int i = 0; i < num_columns; ++i) {
-        if (columns[i] > 1e-6) {
-            if (columns[i] < 1.0 - 1e-8) {
+        if (aux_cols[i] > EPS_SOLVER) {
+            if (aux_cols[i] < 1.0 - EPS_SOLVER) {
                 set_is_integer_solution(false);
             }
 
-            auto counter = 0u;
-            auto tmp = (ScheduleSet*)g_ptr_array_index(schedule_sets, i);
-            auto tmp_nodeid(decision_diagram->root());
+            auto      counter = 0u;
+            auto      tmp = static_cast<ScheduleSet*>(aux_schedule_sets[i]);
+            auto      tmp_nodeid(decision_diagram.root());
+            std::span aux_jobs{tmp->job_list->pdata, tmp->job_list->len};
 
             while (tmp_nodeid > 1) {
                 Job* tmp_j = nullptr;
 
                 if (counter < tmp->job_list->len) {
-                    tmp_j = (Job*)g_ptr_array_index(tmp->job_list, counter);
+                    tmp_j = static_cast<Job*>(aux_jobs[counter]);
                 }
 
                 auto& tmp_node = table.node(tmp_nodeid);
 
                 if (tmp_j == tmp_node.get_job()) {
-                    tmp_node.lp_x[1] += columns[i];
+                    tmp_node.lp_x[1] += aux_cols[i];
                     tmp_nodeid = tmp_node.branch[1];
                     counter++;
                 } else {
-                    tmp_node.lp_x[0] += columns[i];
+                    tmp_node.lp_x[0] += aux_cols[i];
                     tmp_nodeid = tmp_node.branch[0];
                 }
             }
@@ -1497,23 +1470,23 @@ void PricerSolverBdd::construct_lp_sol_from_rmp(const double*    columns,
     }
 
     lp_sol.clear();
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             it.lp_visited = false;
             auto value = it.lp_x[1];
-            if (value > 1e-6) {
-                lp_sol.push_back(
-                    BddCoeff(it.get_nb_job(), it.get_weight(), 0.0, value));
+            if (value > EPS_SOLVER) {
+                lp_sol.emplace_back(it.get_nb_job(), it.get_weight(), 0.0,
+                                    value);
             }
             value = it.lp_x[0];
-            if (value > 1e-6) {
-                lp_sol.push_back(BddCoeff(it.get_nb_job(), it.get_weight(), 0.0,
-                                          value, -1, false));
+            if (value > EPS_SOLVER) {
+                lp_sol.emplace_back(it.get_nb_job(), it.get_weight(), 0.0,
+                                    value, -1, false);
             }
         }
     }
 
-    if (is_integer_solution) {
+    if (is_integer_solution && dbg_lvl() > 1) {
         fmt::print("FOUND INTEGER SOLUTION\n\n");
     }
 
@@ -1526,30 +1499,30 @@ void PricerSolverBdd::construct_lp_sol_from_rmp(const double*    columns,
     // outf.close();
 }
 
-void PricerSolverBdd::calculate_job_time(std::vector<std::vector<double>>& v) {
+void PricerSolverBdd::calculate_job_time(std::vector<std::vector<double>>* v) {
     for (auto& it : lp_sol) {
         if (it.get_high()) {
-            v[it.get_j()][it.get_t()] += it.get_value();
+            (*v)[it.get_j()][it.get_t()] += it.get_value();
         }
     }
 }
 
-void PricerSolverBdd::split_job_time(int _job, int _time, bool _left = false) {
-    auto& table = *(decision_diagram->getDiagram());
+void PricerSolverBdd::split_job_time(int _job, int _time, bool _left) {
+    auto& table = *(decision_diagram.getDiagram());
     auto  removed_edges = false;
 
-    for (auto i = decision_diagram->topLevel(); i > 0; i--) {
+    for (auto i = decision_diagram.topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             if (_left) {
-                if (it.get_weight() + it.get_job()->processing_time > _time &&
+                if (it.get_weight() + it.get_job()->processing_time <= _time &&
                     it.get_nb_job() == _job) {
-                    it.calc_yes = false;
+                    it.calc[1] = false;
                     removed_edges = true;
                 }
             } else {
-                if (it.get_weight() + it.get_job()->processing_time <= _time &&
+                if (it.get_weight() + it.get_job()->processing_time > _time &&
                     it.get_nb_job() == _job) {
-                    it.calc_yes = false;
+                    it.calc[1] = false;
                     removed_edges = true;
                 }
             }
@@ -1565,7 +1538,7 @@ void PricerSolverBdd::split_job_time(int _job, int _time, bool _left = false) {
         construct_mipgraph();
 
         if (dbg_lvl() > 0) {
-            auto&             table_bis = *(decision_diagram->getDiagram());
+            auto&             table_bis = *(decision_diagram.getDiagram());
             ColorWriterVertex vertex_writer(mip_graph, table_bis);
             auto file_name = "split_solution_" + problem_name + "_" +
                              std::to_string(_job) + "_" +
@@ -1579,14 +1552,14 @@ void PricerSolverBdd::split_job_time(int _job, int _time, bool _left = false) {
 }
 
 int PricerSolverBdd::add_constraints() {
-    auto& table = *(decision_diagram->getDiagram());
+    auto& table = *(decision_diagram.getDiagram());
     if (get_is_integer_solution()) {
         bool added_cuts = false;
         return added_cuts;
     } else {
         auto generator =
             ZeroHalfCuts(convex_constr_id, convex_rhs, &reformulation_model,
-                         decision_diagram->root(), &table);
+                         decision_diagram.root(), &table);
 
         generator.generate_cuts();
 
@@ -1608,8 +1581,8 @@ void PricerSolverBdd::remove_constraints(int first, int nb_del) {
 
 void PricerSolverBdd::update_rows_coeff(int first) {
     for (int k = first; k < original_model.get_nb_constraints(); k++) {
-        auto aux = original_model.get_coeff_list(k);
-        for (auto& it : *aux) {
+        auto& aux = original_model.get_coeff_list(k);
+        for (auto& it : aux) {
             it->set_row(k);
         }
     }
@@ -1617,16 +1590,17 @@ void PricerSolverBdd::update_rows_coeff(int first) {
 
 bool PricerSolverBdd::check_schedule_set(GPtrArray* set) {
     // guint              weight = 0;
-    auto& table = *(decision_diagram->getDiagram());
-    auto  tmp_nodeid(decision_diagram->root());
-    auto  counter = 0u;
+    auto&     table = *(decision_diagram.getDiagram());
+    auto      tmp_nodeid(decision_diagram.root());
+    auto      counter = 0u;
+    std::span aux_set{set->pdata, set->len};
 
     while (tmp_nodeid > 1) {
         auto& tmp_node = table.node(tmp_nodeid);
         Job*  tmp_j = nullptr;
 
         if (counter < set->len) {
-            tmp_j = (Job*)g_ptr_array_index(set, counter);
+            tmp_j = static_cast<Job*>(aux_set[counter]);
         }
 
         if (tmp_j == tmp_node.get_job()) {
@@ -1644,10 +1618,10 @@ void PricerSolverBdd::make_schedule_set_feasible(
     [[maybe_unused]] GPtrArray* set) {}
 
 void PricerSolverBdd::iterate_zdd() {
-    DdStructure<NodeBdd<double>>::const_iterator it = decision_diagram->begin();
+    DdStructure<NodeBdd<double>>::const_iterator it = decision_diagram.begin();
 
-    for (; it != decision_diagram->end(); ++it) {
-        std::set<int>::const_iterator i = (*it).begin();
+    for (; it != decision_diagram.end(); ++it) {
+        auto i = (*it).begin();
 
         for (; i != (*it).end(); ++i) {
             fmt::print("{} ", ordered_jobs->len - *i);
@@ -1660,7 +1634,7 @@ void PricerSolverBdd::iterate_zdd() {
 void PricerSolverBdd::create_dot_zdd(const char* name) {
     std::ofstream file;
     file.open(name);
-    decision_diagram->dumpDot(file);
+    // decision_diagram.dumpDot(file);
     file.close();
 }
 
@@ -1686,7 +1660,7 @@ size_t PricerSolverBdd::get_nb_vertices() {
 }
 
 int PricerSolverBdd::get_num_layers() {
-    return decision_diagram->topLevel();
+    return decision_diagram.topLevel();
 }
 
 void PricerSolverBdd::print_num_paths() {}
