@@ -1,4 +1,5 @@
 #include "PricerSolverBddForward.hpp"
+#include <span>
 #include "fmt/core.h"
 
 /**
@@ -7,60 +8,62 @@
 PricerSolverBddSimple::PricerSolverBddSimple(GPtrArray*  _jobs,
                                              int         _num_machines,
                                              GPtrArray*  _ordered_jobs,
-                                             const char* p_name,
-                                             int         _Hmax,
+                                             const char* _p_name,
+                                             int         _hmax,
                                              int*        _take_jobs,
-                                             double      _UB)
+                                             double      _ub)
     : PricerSolverBdd(_jobs,
                       _num_machines,
                       _ordered_jobs,
-                      p_name,
-                      _Hmax,
+                      _p_name,
+                      _hmax,
                       _take_jobs,
-                      _UB) {
-    fmt::print("{0: <{1}}{2}\n", "Constructing BDD with evaluator:", 60,
+                      _ub) {
+    fmt::print("{0: <{1}}{2}\n", "Constructing BDD with evaluator:", ALIGN,
                "Forward Simple Evaluator");
-    fmt::print("{0: <{1}}{2}\n", "Number of vertices BDD", 60,
+    fmt::print("{0: <{1}}{2}\n", "Number of vertices BDD", ALIGN,
                get_nb_vertices());
-    fmt::print("{0: <{1}}{2}\n", "Number of edges BDD", 60, get_nb_edges());
+    fmt::print("{0: <{1}}{2}\n", "Number of edges BDD", ALIGN, get_nb_edges());
 }
 
 OptimalSolution<double> PricerSolverBddSimple::pricing_algorithm(double* _pi) {
     evaluator.set_pi(_pi);
-    return get_decision_diagram()->evaluate_forward(evaluator);
+    return get_decision_diagram().evaluate_forward(evaluator);
 }
 
 OptimalSolution<double> PricerSolverBddSimple::farkas_pricing(double* _pi) {
     farkas_evaluator.set_pi(_pi);
-    return get_decision_diagram()->evaluate_backward(farkas_evaluator);
+    return get_decision_diagram().evaluate_backward(farkas_evaluator);
 }
 
 void PricerSolverBddSimple::compute_labels(double* _pi) {
     evaluator.set_pi(_pi);
     reversed_evaluator.set_pi(_pi);
-    get_decision_diagram()->compute_labels_forward(evaluator);
-    get_decision_diagram()->compute_labels_backward(reversed_evaluator);
+    get_decision_diagram().compute_labels_forward(evaluator);
+    get_decision_diagram().compute_labels_backward(reversed_evaluator);
 }
 
 void PricerSolverBddSimple::evaluate_nodes(double* pi, int UB, double LB) {
-    auto& table = *(get_decision_diagram()->getDiagram());
+    auto& table = *(get_decision_diagram().getDiagram());
     compute_labels(pi);
-    double reduced_cost =
-        table.node(1).forward_label[0].get_f() + pi[convex_constr_id];
+    auto      nb_constraints{reformulation_model.get_nb_constraints()};
+    std::span aux_pi{pi, nb_constraints};
+    double    reduced_cost =
+        table.node(1).forward_label[0].get_f() + aux_pi[convex_constr_id];
     bool removed_edges = false;
     int  nb_removed_edges_evaluate = 0;
 
     /** check for each node the Lagrangian dual */
-    for (int i = get_decision_diagram()->topLevel(); i > 0; i--) {
+    for (int i = get_decision_diagram().topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             auto&  child = table.node(it.branch[1]);
             double result = it.forward_label[0].get_f() +
                             child.backward_label[0].get_f() +
-                            it.reduced_cost[1] + pi[convex_constr_id];
+                            it.reduced_cost[1] + aux_pi[convex_constr_id];
             auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
-            if (LB + aux_nb_machines * reduced_cost + result > UB + 0.0001 &&
-                (it.calc_yes)) {
-                it.calc_yes = false;
+            if (LB + aux_nb_machines * reduced_cost + result > UB + RC_FIXING &&
+                (it.calc[1])) {
+                it.calc[1] = false;
                 add_nb_removed_edges();
                 removed_edges = true;
                 nb_removed_edges_evaluate++;
@@ -70,10 +73,10 @@ void PricerSolverBddSimple::evaluate_nodes(double* pi, int UB, double LB) {
 
     if (removed_edges) {
         fmt::print("Number of edges removed by evaluate nodes {{0}:<{1}}\n",
-                   nb_removed_edges_evaluate, 30);
+                   nb_removed_edges_evaluate, ALIGN_HALF);
         fmt::print("Total number of edges removed {{0}:<{1}}\n",
-                   get_nb_removed_edges(), 30);
-        fmt::print("Number of edges {{0}:<{1}}\n", get_nb_edges(), 30);
+                   get_nb_removed_edges(), ALIGN_HALF);
+        fmt::print("Number of edges {{0}:<{1}}\n", get_nb_edges(), ALIGN_HALF);
         remove_layers();
         remove_edges();
         bottum_up_filtering();
@@ -84,23 +87,24 @@ void PricerSolverBddSimple::evaluate_nodes(double* pi, int UB, double LB) {
 }
 
 void PricerSolverBddSimple::evaluate_nodes(double* pi) {
-    auto& table = *(get_decision_diagram()->getDiagram());
+    auto& table = *(get_decision_diagram().getDiagram());
     compute_labels(pi);
     double reduced_cost = table.node(1).forward_label[0].get_f();
     bool   removed_edges = false;
     int    nb_removed_edges_evaluate = 0;
 
     /** check for each node the Lagrangian dual */
-    for (int i = get_decision_diagram()->topLevel(); i > 0; i--) {
+    for (int i = get_decision_diagram().topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             auto&  child = table.node(it.branch[1]);
             double result = it.forward_label[0].get_f() +
                             child.backward_label[0].get_f() +
                             it.reduced_cost[1];
             auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
-            if (constLB + aux_nb_machines * reduced_cost + result > UB + 1e-4 &&
-                (it.calc_yes)) {
-                it.calc_yes = false;
+            if (constLB + aux_nb_machines * reduced_cost + result >
+                    UB + RC_FIXING &&
+                (it.calc[1])) {
+                it.calc[1] = false;
                 add_nb_removed_edges();
                 removed_edges = true;
                 nb_removed_edges_evaluate++;
@@ -110,10 +114,10 @@ void PricerSolverBddSimple::evaluate_nodes(double* pi) {
 
     if (removed_edges) {
         fmt::print("Number of edges removed by evaluate nodes {0:<{1}}\n",
-                   nb_removed_edges_evaluate, 30);
+                   nb_removed_edges_evaluate, ALIGN_HALF);
         fmt::print("Total number of edges removed {0:<{1}}\n",
-                   get_nb_removed_edges(), 30);
-        fmt::print("Number of edges {0:<{1}}\n", get_nb_edges(), 30);
+                   get_nb_removed_edges(), ALIGN_HALF);
+        fmt::print("Number of edges {0:<{1}}\n", get_nb_edges(), ALIGN_HALF);
         remove_layers();
         remove_edges();
         bottum_up_filtering();
@@ -130,81 +134,83 @@ void PricerSolverBddSimple::evaluate_nodes(double* pi) {
 PricerSolverBddCycle::PricerSolverBddCycle(GPtrArray*  _jobs,
                                            int         _num_machines,
                                            GPtrArray*  _ordered_jobs,
-                                           const char* p_name,
-                                           int         _Hmax,
+                                           const char* _p_name,
+                                           int         _hmax,
                                            int*        _take_jobs,
-                                           double      _UB)
+                                           double      _ub)
     : PricerSolverBdd(_jobs,
                       _num_machines,
                       _ordered_jobs,
-                      p_name,
-                      _Hmax,
+                      _p_name,
+                      _hmax,
                       _take_jobs,
-                      _UB) {
-    fmt::print("{0: <{1}}{2}\n", "Constructing BDD with evaluator:", 60,
+                      _ub) {
+    fmt::print("{0: <{1}}{2}\n", "Constructing BDD with evaluator:", ALIGN,
                "Forward Cycle Evaluator");
-    fmt::print("{0: <{1}}{2}\n", "Number of vertices BDD", 60,
+    fmt::print("{0: <{1}}{2}\n", "Number of vertices BDD", ALIGN,
                get_nb_vertices());
-    fmt::print("{0: <{1}}{2}\n", "Number of edges BDD", 60, get_nb_edges());
+    fmt::print("{0: <{1}}{2}\n", "Number of edges BDD", ALIGN, get_nb_edges());
 }
 
 OptimalSolution<double> PricerSolverBddCycle::pricing_algorithm(double* _pi) {
     evaluator.set_pi(_pi);
-    return get_decision_diagram()->evaluate_forward(evaluator);
+    return get_decision_diagram().evaluate_forward(evaluator);
 }
 
 OptimalSolution<double> PricerSolverBddCycle::farkas_pricing(double* _pi) {
     farkas_evaluator.set_pi(_pi);
-    return get_decision_diagram()->evaluate_backward(farkas_evaluator);
+    return get_decision_diagram().evaluate_backward(farkas_evaluator);
 }
 
 void PricerSolverBddCycle::compute_labels(double* _pi) {
     evaluator.set_pi(_pi);
     reversed_evaluator.set_pi(_pi);
-    get_decision_diagram()->compute_labels_forward(evaluator);
-    get_decision_diagram()->compute_labels_backward(reversed_evaluator);
+    get_decision_diagram().compute_labels_forward(evaluator);
+    get_decision_diagram().compute_labels_backward(reversed_evaluator);
 }
 
 void PricerSolverBddCycle::evaluate_nodes(double* pi, int UB, double LB) {
-    auto& table = *(get_decision_diagram()->getDiagram());
+    auto& table = *(get_decision_diagram().getDiagram());
     compute_labels(pi);
-    double reduced_cost =
-        table.node(1).forward_label[0].get_f() + pi[convex_constr_id];
+    auto      nb_constraints{reformulation_model.get_nb_constraints()};
+    std::span aux_pi{pi, nb_constraints};
+    double    reduced_cost =
+        table.node(1).forward_label[0].get_f() + aux_pi[convex_constr_id];
     bool removed_edges = false;
     int  nb_removed_edges_evaluate = 0;
 
     /** check for each node the Lagrangian dual */
-    for (int i = get_decision_diagram()->topLevel(); i > 0; i--) {
+    for (int i = get_decision_diagram().topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             Job*   job = it.get_job();
-            double result;
+            double result{};
             auto&  child = table.node(it.branch[1]);
 
             if (it.forward_label[0].get_previous_job() != job &&
                 child.backward_label[0].get_prev_job() != job) {
                 result = it.forward_label[0].get_f() +
                          child.backward_label[0].get_f() + it.reduced_cost[1] +
-                         pi[convex_constr_id];
+                         aux_pi[convex_constr_id];
             } else if (it.forward_label[0].get_previous_job() == job &&
                        child.backward_label[0].get_prev_job() != job) {
                 result = it.forward_label[1].get_f() +
                          child.backward_label[0].get_f() + it.reduced_cost[1] +
-                         pi[convex_constr_id];
+                         aux_pi[convex_constr_id];
             } else if (it.forward_label[0].get_previous_job() != job &&
                        child.backward_label[0].get_prev_job() == job) {
                 result = it.forward_label[0].get_f() +
                          child.backward_label[1].get_f() + it.reduced_cost[1] +
-                         pi[convex_constr_id];
+                         aux_pi[convex_constr_id];
             } else {
                 result = it.forward_label[1].get_f() +
                          child.backward_label[1].get_f() + it.reduced_cost[1] +
-                         pi[convex_constr_id];
+                         aux_pi[convex_constr_id];
             }
 
             auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
-            if (LB + aux_nb_machines * reduced_cost + result > UB + 0.0001 &&
-                (it.calc_yes)) {
-                it.calc_yes = false;
+            if (LB + aux_nb_machines * reduced_cost + result > UB + RC_FIXING &&
+                (it.calc[1])) {
+                it.calc[1] = false;
                 removed_edges = true;
                 add_nb_removed_edges();
                 nb_removed_edges_evaluate++;
@@ -271,10 +277,10 @@ void PricerSolverBddCycle::evaluate_nodes(double* pi, int UB, double LB) {
 
     if (removed_edges) {
         fmt::print("Number of edges removed by evaluate nodes {0: <{1}}\n",
-                   nb_removed_edges_evaluate, 30);
+                   nb_removed_edges_evaluate, ALIGN_HALF);
         fmt::print("Total number of edges removed {0: <{1}}\n",
-                   get_nb_removed_edges(), 30);
-        fmt::print("Number of edges {0: <{1}}\n", get_nb_edges(), 30);
+                   get_nb_removed_edges(), ALIGN_HALF);
+        fmt::print("Number of edges {0: <{1}}\n", get_nb_edges(), ALIGN_HALF);
         remove_layers();
         remove_edges();
         bottum_up_filtering();
@@ -285,17 +291,17 @@ void PricerSolverBddCycle::evaluate_nodes(double* pi, int UB, double LB) {
 }
 
 void PricerSolverBddCycle::evaluate_nodes(double* pi) {
-    auto& table = *(get_decision_diagram()->getDiagram());
+    auto& table = *(get_decision_diagram().getDiagram());
     compute_labels(pi);
     double reduced_cost = table.node(1).forward_label[0].get_f();
     bool   removed_edges = false;
     int    nb_removed_edges_evaluate = 0;
 
     /** check for each node the Lagrangian dual */
-    for (int i = get_decision_diagram()->topLevel(); i > 0; i--) {
+    for (int i = get_decision_diagram().topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
             Job*   job = it.get_job();
-            double result;
+            double result{};
             auto&  child = table.node(it.branch[1]);
 
             if (it.forward_label[0].get_previous_job() != job &&
@@ -316,9 +322,10 @@ void PricerSolverBddCycle::evaluate_nodes(double* pi) {
             }
 
             auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
-            if (constLB + aux_nb_machines * reduced_cost + result > UB + 1e-4 &&
-                (it.calc_yes)) {
-                it.calc_yes = false;
+            if (constLB + aux_nb_machines * reduced_cost + result >
+                    UB + RC_FIXING &&
+                (it.calc[1])) {
+                it.calc[1] = false;
                 removed_edges = true;
                 add_nb_removed_edges();
                 nb_removed_edges_evaluate++;
@@ -385,10 +392,10 @@ void PricerSolverBddCycle::evaluate_nodes(double* pi) {
 
     if (removed_edges) {
         fmt::print("Number of edges removed by evaluate nodes {0: <{1}}\n",
-                   nb_removed_edges_evaluate, 30);
+                   nb_removed_edges_evaluate, ALIGN_HALF);
         fmt::print("Total number of edges removed {0: <{1}}\n",
-                   get_nb_removed_edges(), 30);
-        fmt::print("Number of edges {0: <{1}}\n", get_nb_edges(), 30);
+                   get_nb_removed_edges(), ALIGN_HALF);
+        fmt::print("Number of edges {0: <{1}}\n", get_nb_edges(), ALIGN_HALF);
         remove_layers();
         remove_edges();
         bottum_up_filtering();
