@@ -4,38 +4,38 @@
 #include <boost/graph/adjacency_list.hpp>
 #include <boost/graph/graph_traits.hpp>
 #include "boost/graph/graphviz.hpp"
+#include "gurobi_c++.h"
+#include "gurobi_c.h"
 
 /**
  * Pricersolver for the TI index formulation
  */
 PricerSolverSimpleDp::PricerSolverSimpleDp(GPtrArray*  _jobs,
                                            int         _num_machines,
-                                           int         _Hmax,
-                                           const char* p_name,
-                                           double      _UB)
-    : PricerSolverBase(_jobs, _num_machines, p_name, _UB),
-      Hmax(_Hmax),
+                                           int         _hmax,
+                                           const char* _p_name,
+                                           double      _ub)
+    : PricerSolverBase(_jobs, _num_machines, _p_name, _ub),
+      Hmax(_hmax),
       size_graph(0u),
-      A(new Job*[Hmax + 1]),
-      F(new double[Hmax + 1]),
-      backward_F(new double[Hmax + 1]),
-      TI_x(new GRBVar[convex_constr_id * (Hmax + 1)]),
-      take(static_cast<int*>(
-          malloc(convex_constr_id * (Hmax + 1) * sizeof(int)))),
-      lp_x(new double[convex_constr_id * (Hmax + 1)]{}),
-      solution_x(new double[convex_constr_id * (Hmax + 1)]{}) {
-    fill_int(take, convex_constr_id * (Hmax + 1), 0);
+      A(Hmax + 1),
+      F(Hmax + 1),
+      backward_F(Hmax + 1),
+      TI_x(convex_constr_id * (Hmax + 1), GRBVar()),
+      take(convex_constr_id * (Hmax + 1), 0),
+      lp_x(convex_constr_id * (Hmax + 1), 0.0),
+      solution_x(convex_constr_id * (Hmax + 1), 0.0) {
     init_table();
 }
 
 void PricerSolverSimpleDp::init_table() {
-    backward_graph = new std::vector<Job*>[Hmax + 1];
-    forward_graph = new std::vector<Job*>[Hmax + 1];
+    backward_graph = std::vector<std::vector<Job*>>(Hmax + 1);
+    forward_graph = std::vector<std::vector<Job*>>(Hmax + 1);
 
     for (int t = 0; t < Hmax + 1; t++) {
         for (int i = 1; i < convex_constr_id + 1; i++) {
             int  j = i - 1;
-            Job* job = reinterpret_cast<Job*>(g_ptr_array_index(jobs, j));
+            Job* job = static_cast<Job*>(jobs[j]);
 
             if (t >= job->processing_time) {
                 forward_graph[t].push_back(job);
@@ -52,14 +52,14 @@ void PricerSolverSimpleDp::init_table() {
 }
 
 PricerSolverSimpleDp::~PricerSolverSimpleDp() {
-    delete[] backward_graph;
-    delete[] forward_graph;
-    delete[] TI_x;
-    if (take) {
-        free(take);
-    }
-    delete[] lp_x;
-    delete[] solution_x;
+    // delete[] backward_graph;
+    // delete[] forward_graph;
+    // delete[] TI_x;
+    // if (take) {
+    //     free(take);
+    // }
+    // delete[] lp_x;
+    // delete[] solution_x;
 }
 
 void PricerSolverSimpleDp::evaluate_nodes(double*                 pi,
@@ -77,15 +77,17 @@ void PricerSolverSimpleDp::evaluate_nodes([[maybe_unused]] double* pi) {
 
 void PricerSolverSimpleDp::reduce_cost_fixing(double* pi, int UB, double LB) {
     evaluate_nodes(pi, UB, LB);
-    int counter = 0;
-    int x = 0;
+    auto      nb_constraints = reformulation_model.get_nb_constraints();
+    std::span aux_pi{pi, nb_constraints};
+    int       counter = 0;
+    int       x = 0;
 
     for (int t = 0; t < Hmax + 1; t++) {
         auto it = forward_graph[t].begin();
         while (it != forward_graph[t].end()) {
             double result = F[t - (*it)->processing_time] + value_Fj(t, *it) -
-                            pi[(*it)->job] + backward_F[t];
-            if (LB + result + (convex_rhs - 1) * F[Hmax] > UB + 1e-6) {
+                            aux_pi[(*it)->job] + backward_F[t];
+            if (LB + result + (convex_rhs - 1) * F[Hmax] > UB + RC_FIXING) {
                 size_graph--;
                 it = forward_graph[t].erase(it);
             } else {
@@ -99,8 +101,8 @@ void PricerSolverSimpleDp::reduce_cost_fixing(double* pi, int UB, double LB) {
         while (iter != backward_graph[t].end()) {
             double result =
                 F[t] + value_Fj(t + (*iter)->processing_time, *iter) -
-                pi[(*iter)->job] + backward_F[t + (*iter)->processing_time];
-            if (LB + result + (convex_rhs - 1) * F[Hmax] > UB + 1e-6) {
+                aux_pi[(*iter)->job] + backward_F[t + (*iter)->processing_time];
+            if (LB + result + (convex_rhs - 1) * F[Hmax] > UB + RC_FIXING) {
                 take[(*iter)->job * (Hmax + 1) + t] = false;
                 // iter = backward_graph[t].erase(iter);
             } else {
@@ -126,22 +128,16 @@ void PricerSolverSimpleDp::build_mip() {
                 double cost = value_Fj(t + it->processing_time, it);
                 double ub = take[(it->job) * (Hmax + 1) + t] ? 1.0 : 0.0;
                 TI_x[it->job * (Hmax + 1) + t] =
-                    model->addVar(0.0, ub, cost, GRB_BINARY);
+                    model.addVar(0.0, ub, cost, GRB_BINARY);
             }
         }
 
-        model->update();
+        model.update();
 
         /** Assignment variables */
-        std::unique_ptr<GRBLinExpr[]> assignment(
-            new GRBLinExpr[convex_constr_id]());
-        std::unique_ptr<char[]>   sense(new char[convex_constr_id]);
-        std::unique_ptr<double[]> rhs(new double[convex_constr_id]);
-
-        for (unsigned i = 0; i < jobs->len; ++i) {
-            sense[i] = GRB_EQUAL;
-            rhs[i] = 1.0;
-        }
+        std::vector<GRBLinExpr> assignment(convex_constr_id, GRBLinExpr());
+        std::vector<char>       sense(convex_constr_id, GRB_EQUAL);
+        std::vector<double>     rhs(convex_constr_id, 1.0);
 
         for (int t = 0; t <= Hmax; t++) {
             for (auto& it : backward_graph[t]) {
@@ -149,16 +145,15 @@ void PricerSolverSimpleDp::build_mip() {
             }
         }
 
-        std::unique_ptr<GRBConstr[]> assignment_constrs(
-            model->addConstrs(assignment.get(), sense.get(), rhs.get(), nullptr,
-                              convex_constr_id));
+        std::unique_ptr<GRBConstr> assignment_constrs(
+            model.addConstrs(assignment.data(), sense.data(), rhs.data(),
+                             nullptr, convex_constr_id));
 
-        model->update();
+        model.update();
 
-        std::unique_ptr<GRBLinExpr[]> interval_constr(
-            new GRBLinExpr[Hmax + 1]());
-        std::unique_ptr<char[]>   interval_sense(new char[Hmax + 1]);
-        std::unique_ptr<double[]> interval_rhs(new double[Hmax + 1]);
+        std::vector<GRBLinExpr> interval_constr(Hmax + 1, GRBLinExpr());
+        std::vector<char>       interval_sense(Hmax + 1);
+        std::vector<double>     interval_rhs(Hmax + 1);
 
         for (int t = 0; t <= Hmax; t++) {
             auto add_constraint = false;
@@ -178,12 +173,12 @@ void PricerSolverSimpleDp::build_mip() {
                 interval_sense[t] = GRB_LESS_EQUAL;
                 interval_rhs[t] = convex_rhs;
 
-                model->addConstr(interval_constr[t], interval_sense[t],
-                                 interval_rhs[t]);
+                model.addConstr(interval_constr[t], interval_sense[t],
+                                interval_rhs[t]);
             }
         }
 
-        model->update();
+        model.update();
 
     } catch (GRBException e) {
         std::cerr << e.getMessage() << '\n';
@@ -196,15 +191,17 @@ void PricerSolverSimpleDp::build_mip() {
         }
     }
 
-    model->write("ti_" + problem_name + "_" + std::to_string(convex_rhs) +
-                 "correct.lp");
-    model->optimize();
+    model.write("ti_" + problem_name + "_" + std::to_string(convex_rhs) +
+                "correct.lp");
+    model.optimize();
     return;
 }
 
 void PricerSolverSimpleDp::forward_evaluator(double* _pi) {
     /** Initialisation */
-    F[0] = _pi[convex_constr_id];
+    auto      nb_constraints = reformulation_model.get_nb_constraints();
+    std::span aux_pi{_pi, nb_constraints};
+    F[0] = aux_pi[convex_constr_id];
     A[0] = nullptr;
 
     for (int t = 1; t < Hmax + 1; t++) {
@@ -216,10 +213,10 @@ void PricerSolverSimpleDp::forward_evaluator(double* _pi) {
     for (int t = 1; t < Hmax + 1; t++) {
         for (auto& it : forward_graph[t]) {
             if (F[t - it->processing_time] +
-                    static_cast<double>(value_Fj(t, it)) - _pi[it->job] <=
+                    static_cast<double>(value_Fj(t, it)) - aux_pi[it->job] <=
                 F[t]) {
-                F[t] =
-                    F[t - it->processing_time] + value_Fj(t, it) - _pi[it->job];
+                F[t] = F[t - it->processing_time] + value_Fj(t, it) -
+                       aux_pi[it->job];
                 A[t] = it;
             }
         }
@@ -232,6 +229,8 @@ void PricerSolverSimpleDp::forward_evaluator(double* _pi) {
 
 void PricerSolverSimpleDp::backward_evaluator(double* _pi) {
     backward_F[Hmax] = 0.0;
+    auto      nb_constraints{reformulation_model.get_nb_constraints()};
+    std::span aux_pi{_pi, nb_constraints};
 
     for (int t = 0; t < Hmax; t++) {
         backward_F[t] = DBL_MAX / 2;
@@ -242,11 +241,11 @@ void PricerSolverSimpleDp::backward_evaluator(double* _pi) {
             Job* job = it;
             int  tt = t + job->processing_time;
             if (backward_F[tt] + static_cast<double>(value_Fj(tt, job)) -
-                    _pi[job->job] <=
+                    aux_pi[job->job] <=
                 backward_F[t]) {
                 backward_F[t] = backward_F[tt] +
                                 static_cast<double>(value_Fj(tt, job)) -
-                                _pi[job->job];
+                                aux_pi[job->job];
             }
         }
 
@@ -284,7 +283,7 @@ OptimalSolution<double> PricerSolverSimpleDp::pricing_algorithm(double* _pi) {
         t_min -= job->processing_time;
     }
 
-    std::vector<Job*>::reverse_iterator it = v.rbegin();
+    auto it = v.rbegin();
 
     for (; it != v.rend(); ++it) {
         g_ptr_array_add(opt_sol.jobs, *it);
@@ -305,15 +304,18 @@ void PricerSolverSimpleDp::construct_lp_sol_from_rmp(
     const double*    columns,
     const GPtrArray* schedule_sets,
     int              num_columns) {
-    std::fill(lp_x, lp_x + convex_constr_id * (Hmax + 1), 0.0);
+    auto      nb_constraints{reformulation_model.get_nb_constraints()};
+    std::span aux_cols{columns, schedule_sets->len};
+    std::span aux_schedule_sets{schedule_sets->pdata, schedule_sets->len};
+    std::fill(lp_x.begin(), lp_x.end(), 0.0);
     for (int k = 0; k < num_columns; k++) {
-        if (columns[k] > 0.00001) {
-            ScheduleSet* tmp =
-                (ScheduleSet*)g_ptr_array_index(schedule_sets, k);
-            int t = 0;
-            for (size_t l = 0; l < tmp->job_list->len; l++) {
-                Job* tmp_j = (Job*)g_ptr_array_index(tmp->job_list, l);
-                lp_x[(tmp_j->job) * (Hmax + 1) + t] += columns[k];
+        if (aux_cols[k] > EPS_SOLVER) {
+            auto*     tmp = static_cast<ScheduleSet*>(aux_schedule_sets[k]);
+            int       t = 0;
+            std::span aux_jobs{tmp->job_list->pdata, tmp->job_list->len};
+            for (auto& it : aux_jobs) {
+                Job* tmp_j = static_cast<Job*>(it);
+                lp_x[(tmp_j->job) * (Hmax + 1) + t] += aux_cols[k];
                 t += tmp_j->processing_time;
             }
         }
@@ -382,8 +384,8 @@ void PricerSolverSimpleDp::print_num_paths() {}
 bool PricerSolverSimpleDp::check_schedule_set([[maybe_unused]] GPtrArray* set) {
     // int t = 0;
     // for(unsigned int j = 0; j < set->len; j++) {
-    //     Job* tmp_j = (Job*) g_ptr_array_index(set, j);
-    //     t += tmp_j->processing_time;
+    //     Job* tmp_j = static_cast<Job*>() g_ptr_array_index()atic_cast<set,
+    //     j>())); t += tmp_j->processing_time;
 
     //     if (t > Hmax) {
     //         return false;
