@@ -8,9 +8,9 @@ PricerSolverZddBackwardSimple::PricerSolverZddBackwardSimple(
     GPtrArray*  _jobs,
     int         _num_machines,
     GPtrArray*  _ordered_jobs,
-    const char* p_name,
-    double      _UB)
-    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, p_name, _UB) {
+    const char* _p_name,
+    double      _ub)
+    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, _p_name, _ub) {
     std::cout << "Constructing ZDD with Backward Simple evaluator" << '\n';
     std::cout << "number vertices ZDD = " << get_nb_vertices() << '\n';
     std::cout << "number edges ZDD = " << get_nb_edges() << '\n';
@@ -54,7 +54,7 @@ void PricerSolverZddBackwardSimple::evaluate_nodes(double* pi,
                                 pi[job->job] + pi[convex_constr_id];
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
                 if (LB - aux_nb_machines * reduced_cost - result >
-                        UB - 1 + 0.0001 &&
+                        UB - 1 + RC_FIXING &&
                     (iter->calc_yes)) {
                     iter->calc_yes = false;
                     nb_removed_edges++;
@@ -86,7 +86,7 @@ void PricerSolverZddBackwardSimple::evaluate_nodes(double* pi) {
                                 pi[job->job];
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
                 if (constLB + aux_nb_machines * reduced_cost + result >
-                        UB - 1 + 0.0001 &&
+                        UB - 1 + RC_FIXING &&
                     (iter->calc_yes)) {
                     iter->calc_yes = false;
                     nb_removed_edges++;
@@ -102,9 +102,9 @@ PricerSolverZddBackwardCycle::PricerSolverZddBackwardCycle(
     GPtrArray*  _jobs,
     int         _num_machines,
     GPtrArray*  _ordered_jobs,
-    const char* p_name,
-    double      _UB)
-    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, p_name, _UB) {
+    const char* _p_name,
+    double      _ub)
+    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, _p_name, _ub) {
     std::cout << "Constructing ZDD with Backward ZddCycle evaluator" << '\n';
     std::cout << "number vertices ZDD = " << get_nb_vertices() << '\n';
     std::cout << "number edges ZDD = " << get_nb_edges() << '\n';
@@ -136,19 +136,19 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi) {
     /** check for each node the Lagrangian dual */
     for (int i = decision_diagram->topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
+            auto* job = it.get_job();
+            auto  p = job->processing_time;
             for (auto& iter : it.list) {
-                int  w = iter->get_weight();
-                Job* job = it.get_job();
+                auto w = iter->get_weight();
 
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
                 if (iter->forward_label[0].get_previous_job() != job &&
                     iter->y->backward_label[0].get_prev_job() != job) {
                     double result = iter->forward_label[0].get_f() +
                                     iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
+                                    value_Fj(w + p, job) + pi[job->job];
                     if (constLB + aux_nb_machines * reduced_cost + result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -157,10 +157,9 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi) {
                            iter->y->backward_label[0].get_prev_job() != job) {
                     double result = iter->forward_label[1].get_f() +
                                     iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
+                                    value_Fj(w + p, job) + pi[job->job];
                     if (constLB + aux_nb_machines * reduced_cost + result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -169,10 +168,9 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi) {
                            iter->y->backward_label[0].get_prev_job() == job) {
                     double result = iter->forward_label[0].get_f() +
                                     iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
+                                    value_Fj(w + p, job) + pi[job->job];
                     if (constLB + aux_nb_machines * reduced_cost + result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -180,10 +178,9 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi) {
                 } else {
                     double result = iter->forward_label[1].get_f() +
                                     iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
+                                    value_Fj(w + p, job) + pi[job->job];
                     if (constLB + aux_nb_machines * reduced_cost + result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -208,19 +205,20 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi,
     /** check for each node the Lagrangian dual */
     for (int i = decision_diagram->topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
+            auto* job = it.get_job();
+            auto  p = job->processing_time;
             for (auto& iter : it.list) {
-                int  w = iter->get_weight();
-                Job* job = it.get_job();
+                int w = iter->get_weight();
 
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
                 if (iter->forward_label[0].get_previous_job() != job &&
                     iter->y->backward_label[0].get_prev_job() != job) {
                     double result = iter->forward_label[0].get_f() +
                                     iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
+                                    value_Fj(w + p, job) + pi[job->job] +
+                                    pi[convex_constr_id];
                     if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -229,10 +227,10 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi,
                            iter->y->backward_label[0].get_prev_job() != job) {
                     double result = iter->forward_label[1].get_f() +
                                     iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
+                                    value_Fj(w + p, job) + pi[job->job] +
+                                    pi[convex_constr_id];
                     if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -241,10 +239,10 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi,
                            iter->y->backward_label[0].get_prev_job() == job) {
                     double result = iter->forward_label[0].get_f() +
                                     iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
+                                    value_Fj(w + p, job) + pi[job->job] +
+                                    pi[convex_constr_id];
                     if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
@@ -252,10 +250,10 @@ void PricerSolverZddBackwardCycle::evaluate_nodes(double* pi,
                 } else {
                     double result = iter->forward_label[1].get_f() +
                                     iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
+                                    value_Fj(w + p, job) + pi[job->job] +
+                                    pi[convex_constr_id];
                     if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
+                            UB + RC_FIXING &&
                         (iter->calc_yes)) {
                         iter->calc_yes = false;
                         nb_removed_edges++;
