@@ -6,9 +6,9 @@
 PricerSolverSimple::PricerSolverSimple(GPtrArray*  _jobs,
                                        int         _num_machines,
                                        GPtrArray*  _ordered_jobs,
-                                       const char* p_name,
-                                       double      _UB)
-    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, p_name, _UB) {
+                                       const char* _p_name,
+                                       double      _ub)
+    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, _p_name, _ub) {
     std::cout << "Constructing ZDD with Forward Simple evaluator" << '\n';
     std::cout << "number vertices ZDD = " << get_nb_vertices() << '\n';
     std::cout << "number edges ZDD = " << get_nb_edges() << '\n';
@@ -49,7 +49,7 @@ void PricerSolverSimple::evaluate_nodes(double* pi, int UB, double LB) {
                                 pi[job->job] + pi[convex_constr_id];
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
                 if (LB - aux_nb_machines * reduced_cost - result >
-                        UB - 1 + 0.0001 &&
+                        UB - 1 + RC_FIXING &&
                     (iter->calc_yes)) {
                     iter->calc_yes = false;
                     nb_removed_edges++;
@@ -81,7 +81,7 @@ void PricerSolverSimple::evaluate_nodes(double* pi) {
                                 pi[job->job];
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
                 if (constLB + aux_nb_machines * reduced_cost + result >
-                        UB - 1 + 1e-4 &&
+                        UB - 1 + RC_FIXING &&
                     (iter->calc_yes)) {
                     iter->calc_yes = false;
                     nb_removed_edges++;
@@ -96,9 +96,9 @@ void PricerSolverSimple::evaluate_nodes(double* pi) {
 PricerSolverZddCycle::PricerSolverZddCycle(GPtrArray*  _jobs,
                                            int         _num_machines,
                                            GPtrArray*  _ordered_jobs,
-                                           const char* p_name,
-                                           double      _UB)
-    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, p_name, _UB) {
+                                           const char* _p_name,
+                                           double      _ub)
+    : PricerSolverZdd(_jobs, _num_machines, _ordered_jobs, _p_name, _ub) {
     std::cout << "Constructing ZDD with Forward ZddCycle evaluator" << '\n';
     std::cout << "number vertices ZDD = " << get_nb_vertices() << '\n';
     std::cout << "number edges ZDD = " << get_nb_edges() << '\n';
@@ -129,127 +129,127 @@ void PricerSolverZddCycle::evaluate_nodes(double* pi, int UB, double LB) {
     /** check for each node the Lagrangian dual */
     for (int i = decision_diagram->topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
+            auto* job = it.get_job();
             for (auto& iter : it.list) {
-                int  w = iter->get_weight();
-                Job* job = it.get_job();
+                auto w = iter->get_weight();
+                auto p = job->processing_time;
 
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
-                if (iter->forward_label[0].get_previous_job() != job &&
-                    iter->y->backward_label[0].get_prev_job() != job) {
-                    double result = iter->forward_label[0].get_f() +
-                                    iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
-                    if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
-                    }
-                } else if (iter->forward_label[0].get_previous_job() == job &&
-                           iter->y->backward_label[0].get_prev_job() != job) {
-                    double result = iter->forward_label[1].get_f() +
-                                    iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
-                    if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
-                    }
-                } else if (iter->forward_label[0].get_previous_job() != job &&
-                           iter->y->backward_label[0].get_prev_job() == job) {
-                    double result = iter->forward_label[0].get_f() +
-                                    iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
-                    if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
+                if (iter->forward_label[0].get_previous_job() != job) {
+                    if (iter->y->backward_label[0].get_prev_job() != job) {
+                        auto result = iter->forward_label[0].get_f() +
+                                      iter->y->backward_label[0].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job] +
+                                      pi[convex_constr_id];
+                        if (LB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
+                    } else {
+                        auto result = iter->forward_label[0].get_f() +
+                                      iter->y->backward_label[1].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job] +
+                                      pi[convex_constr_id];
+                        if (LB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
                     }
                 } else {
-                    double result = iter->forward_label[1].get_f() +
-                                    iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job] + pi[convex_constr_id];
-                    if (LB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
+                    if (iter->y->backward_label[0].get_prev_job() != job) {
+                        auto result = iter->forward_label[1].get_f() +
+                                      iter->y->backward_label[0].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job] +
+                                      pi[convex_constr_id];
+                        if (LB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
+                    } else {
+                        auto result = iter->forward_label[1].get_f() +
+                                      iter->y->backward_label[1].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job] +
+                                      pi[convex_constr_id];
+                        if (LB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
                     }
                 }
             }
+
+            fmt::print("removed edges = {}\n", nb_removed_edges);
         }
     }
-
-    fmt::print("removed edges = {}\n", nb_removed_edges);
 }
 
 void PricerSolverZddCycle::evaluate_nodes(double* pi) {
     auto& table = *(decision_diagram->getDiagram());
     compute_labels(pi);
-    double reduced_cost =
+    auto reduced_cost =
         table.node(decision_diagram->root()).list[0]->backward_label[0].get_f();
     nb_removed_edges = 0;
 
     /** check for each node the Lagrangian dual */
     for (int i = decision_diagram->topLevel(); i > 0; i--) {
         for (auto& it : table[i]) {
+            auto* job = it.get_job();
             for (auto& iter : it.list) {
-                int  w = iter->get_weight();
-                Job* job = it.get_job();
+                auto w = iter->get_weight();
+                auto p = job->processing_time;
 
                 auto aux_nb_machines = static_cast<double>(convex_rhs - 1);
-                if (iter->forward_label[0].get_previous_job() != job &&
-                    iter->y->backward_label[0].get_prev_job() != job) {
-                    double result = iter->forward_label[0].get_f() +
-                                    iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
-                    if (constLB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
-                    }
-                } else if (iter->forward_label[0].get_previous_job() == job &&
-                           iter->y->backward_label[0].get_prev_job() != job) {
-                    double result = iter->forward_label[1].get_f() +
-                                    iter->y->backward_label[0].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
-                    if (constLB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
-                    }
-                } else if (iter->forward_label[0].get_previous_job() != job &&
-                           iter->y->backward_label[0].get_prev_job() == job) {
-                    double result = iter->forward_label[0].get_f() +
-                                    iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
-                    if (constLB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
+                if (iter->forward_label[0].get_previous_job() != job) {
+                    if (iter->y->backward_label[0].get_prev_job() != job) {
+                        auto result = iter->forward_label[0].get_f() +
+                                      iter->y->backward_label[0].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job];
+                        if (constLB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
+                    } else {
+                        auto result = iter->forward_label[0].get_f() +
+                                      iter->y->backward_label[1].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job];
+                        if (constLB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
                     }
                 } else {
-                    double result = iter->forward_label[1].get_f() +
-                                    iter->y->backward_label[1].get_f() -
-                                    value_Fj(w + job->processing_time, job) +
-                                    pi[job->job];
-                    if (constLB - aux_nb_machines * reduced_cost - result >
-                            UB + 0.0001 &&
-                        (iter->calc_yes)) {
-                        iter->calc_yes = false;
-                        nb_removed_edges++;
+                    if (iter->y->backward_label[0].get_prev_job() != job) {
+                        auto result = iter->forward_label[1].get_f() +
+                                      iter->y->backward_label[0].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job];
+                        if (constLB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
+                    } else {
+                        auto result = iter->forward_label[1].get_f() +
+                                      iter->y->backward_label[1].get_f() -
+                                      value_Fj(w + p, job) + pi[job->job];
+                        if (constLB - aux_nb_machines * reduced_cost - result >
+                                UB + RC_FIXING &&
+                            (iter->calc_yes)) {
+                            iter->calc_yes = false;
+                            nb_removed_edges++;
+                        }
                     }
                 }
             }
