@@ -12,7 +12,7 @@
 ZeroHalfCuts::ZeroHalfCuts(int                 _nb_jobs,
                            int                 _nb_machines,
                            ReformulationModel* _rmp_model,
-                           NodeId&             _root,
+                           NodeId const&       _root,
                            NodeTableEntity<>*  _table)
     : env(std::make_unique<GRBEnv>()),
       model(std::make_unique<GRBModel>(*env)),
@@ -23,22 +23,21 @@ ZeroHalfCuts::ZeroHalfCuts(int                 _nb_jobs,
       table(_table),
       jobs_var(nb_jobs) {
     // Limit how many solutions to collect
-    model->set(GRB_IntParam_PoolSolutions, 124);
+    // model->set(GRB_IntParam_PoolSolutions, 124);
 
     // Limit the search space by setting a gap for the worst possible solution
     // that will be accepted
-    model->set(GRB_DoubleParam_PoolGap, 0.10);
+    // model->set(GRB_DoubleParam_PoolGap, 0.10);
 
     // do a systematic search for the k-best solutions
-    model->set(GRB_IntParam_PoolSearchMode, 2);
+    model->set(GRB_IntParam_PoolSearchMode, GRB_PRESOLVE_AGGRESSIVE);
     model->set(GRB_IntParam_Method, GRB_METHOD_AUTO);
     model->set(GRB_IntParam_Threads, 1);
     model->set(GRB_IntAttr_ModelSense, GRB_MINIMIZE);
-    model->set(GRB_IntParam_Presolve, 2);
-    model->set(GRB_DoubleParam_Heuristics, 0.5);
+    model->set(GRB_IntParam_Presolve, GRB_PRESOLVE_AGGRESSIVE);
+    model->set(GRB_DoubleParam_Heuristics, 1 / HALF);
     model->set(GRB_IntParam_MIPFocus, 1);
-    model->set(GRB_DoubleParam_TimeLimit, 50);
-    model->set(GRB_IntParam_OutputFlag, 1);
+    model->set(GRB_DoubleParam_TimeLimit, TIMELIMIT);
 
     generate_model();
 }
@@ -69,26 +68,25 @@ void ZeroHalfCuts::generate_model() {
         q = model->addVar(0.0, GRB_INFINITY, 0.0, 'I');
 
         expr.addTerms(coeffs.data(), jobs_var.data(), coeffs.size());
-        expr += m * root_node.sigma - m * terminal_node.sigma - 2.0 * q;
+        expr += m * root_node.sigma - m * terminal_node.sigma - HALF * q;
         model->addConstr(expr, '=', 1.0);
         model->update();
         init_table();
 
     } catch (GRBException& e) {
-        fmt::print("Error code = {0} {1:-^{2}} {3}", e.getErrorCode(), "", 40,
-                   e.getMessage());
+        fmt::print("Error code = {0} {1:-^{2}} {3}", e.getErrorCode(), "",
+                   ALIGN, e.getMessage());
     }
 }
 
 void ZeroHalfCuts::init_table() {
     for (auto i = root.row(); i >= 0; i--) {
         for (auto& it : (*table)[i]) {
-            it.in_degree_0 = 0;
-            it.in_degree_1 = 0;
+            it.in_degree = {};
+            // it.in_degree_1 = 0;
             it.in_edges[0].clear();
             it.in_edges[1].clear();
-            it.coeff_cut[0] = 0.0;
-            it.coeff_cut[1] = 0.0;
+            it.coeff_cut = {};
         }
     }
 
@@ -98,9 +96,9 @@ void ZeroHalfCuts::init_table() {
             auto& n1 = table->node(it.branch[1]);
 
             n0.in_edges[0].push_back(it.ptr_node_id);
-            n0.in_degree_0++;
+            n0.in_degree[0]++;
             n1.in_edges[1].push_back(it.ptr_node_id);
-            n1.in_degree_1++;
+            n1.in_degree[1]++;
         }
     }
 }
@@ -108,46 +106,48 @@ void ZeroHalfCuts::init_table() {
 void ZeroHalfCuts::init_coeff_cut() {
     for (auto& it : node_ids) {
         auto& node = table->node(it);
-        init_coeff_node(node);
+        init_coeff_node(&node);
     }
 
     for (auto& iter : node_ids_lift) {
         auto& node = table->node(iter);
-        init_coeff_node(node);
+        init_coeff_node(&node);
     }
     node_ids_lift.clear();
 }
 
-void ZeroHalfCuts::init_coeff_node(NodeBdd<>& node) {
+void ZeroHalfCuts::init_coeff_node(NodeBdd<>* node) {
     for (int k = 0; k < 2; k++) {
-        node.coeff_cut[k] = 0.0;
-        for (auto& it : node.in_edges[k]) {
+        node->coeff_cut.at(k) = 0.0;
+        for (auto& it : node->in_edges.at(k)) {
             auto aux = it.lock();
             if (aux) {
                 auto& aux_node = table->node(*aux);
-                aux_node.coeff_cut[k] = 0.0;
+                aux_node.coeff_cut.at(k) = 0.0;
             }
         }
     }
 }
 
 void ZeroHalfCuts::construct_cut() {
-    GenericData* data = new GenericData();
+    auto* data = new GenericData();
 
     auto add_coeff_constr = [&](const auto& it) {
         auto& node = table->node(it);
 
         for (int k = 0; k < 2; k++) {
-            auto coeff = floor(node.coeff_cut[k] / 2.0);
-            if (coeff > 1e-4)
+            auto coeff = floor(node.coeff_cut.at(k) / HALF);
+            if (coeff > EPS_CUT) {
                 data->add_coeff_hash_table(node.get_nb_job(), node.get_weight(),
                                            k, -coeff);
-            for (auto& iter : node.in_edges[k]) {
+            }
+
+            for (auto& iter : node.in_edges.at(k)) {
                 auto aux = iter.lock();
                 if (aux) {
                     auto& aux_node = table->node(*aux);
-                    auto  coeff_in = floor(aux_node.coeff_cut[k] / 2);
-                    if (coeff_in < -1e-4) {
+                    auto  coeff_in = floor(aux_node.coeff_cut.at(k) / 2);
+                    if (coeff_in < -EPS_CUT) {
                         data->add_coeff_hash_table(aux_node.get_nb_job(),
                                                    aux_node.get_weight(), k,
                                                    -coeff_in);
@@ -157,16 +157,16 @@ void ZeroHalfCuts::construct_cut() {
         }
     };
 
-    auto print_node_ids = [&](const auto& it) {
-        auto& node = table->node(it);
-        if (it > 1 && node.sigma.get(GRB_DoubleAttr_Xn) > 0.0) {
-            fmt::print("node {} {} {} {} {} | ", node.get_nb_job(),
-                       node.get_weight(), node.sigma.get(GRB_DoubleAttr_Xn),
-                       node.lp_x[0], node.lp_x[1]);
-        } else if (node.sigma.get(GRB_DoubleAttr_Xn) > 0.0) {
-            fmt::print("Terminal Node | ");
-        }
-    };
+    // auto print_node_ids = [&](const auto& it) {
+    //     auto& node = table->node(it);
+    //     if (it > 1 && node.sigma.get(GRB_DoubleAttr_Xn) > 0.0) {
+    //         fmt::print("node {} {} {} {} {} | ", node.get_nb_job(),
+    //                    node.get_weight(), node.sigma.get(GRB_DoubleAttr_Xn),
+    //                    node.lp_x[0], node.lp_x[1]);
+    //     } else if (node.sigma.get(GRB_DoubleAttr_Xn) > 0.0) {
+    //         fmt::print("Terminal Node | ");
+    //     }
+    // };
 
     std::for_each(node_ids.begin(), node_ids.end(), add_coeff_constr);
     // std::for_each(node_ids_lift.begin(), node_ids_lift.end(),
@@ -176,22 +176,22 @@ void ZeroHalfCuts::construct_cut() {
     auto& root_node = table->node(root);
     auto& terminal_node = table->node(1);
     auto  rhs = 0.0;
-    rhs += (root_node.sigma.get(GRB_DoubleAttr_Xn) > 1e-6)
+    rhs += (root_node.sigma.get(GRB_DoubleAttr_Xn) > EPS_CUT)
                ? static_cast<double>(nb_machines)
                : 0.0;
-    rhs += (terminal_node.sigma.get(GRB_DoubleAttr_Xn) > 1e-6)
+    rhs += (terminal_node.sigma.get(GRB_DoubleAttr_Xn) > EPS_CUT)
                ? -static_cast<double>(nb_machines)
                : 0.0;
 
     for (auto& it : jobs_var) {
         auto x = it.get(GRB_DoubleAttr_Xn);
-        if (x > 1e-6) {
+        if (x > EPS_CUT) {
             rhs += 1.0;
         }
     }
     // fmt::print("test rhs = {}\n", rhs);
     std::shared_ptr<ConstraintGeneric> constr{
-        std::make_shared<ConstraintGeneric>(data, -floor(rhs / 2.0))};
+        std::make_shared<ConstraintGeneric>(data, -floor(rhs / HALF))};
     // data->list_coeff();
     // fmt::print("RHS = {}\n",
     //            -floor((2.0 * q.get(GRB_DoubleAttr_Xn) + 1 + rhs) / 2.0));
@@ -231,13 +231,12 @@ void ZeroHalfCuts::generate_cuts() {
         for (auto i = 0; i < nb_solutions; i++) {
             model->set(GRB_IntParam_SolutionNumber, i);
             init_coeff_cut();
-            auto sol = model->get(GRB_DoubleAttr_ObjVal);
 
             bool add = false;
             auto calc_coeff_cut = [&](const auto& it) {
                 auto& node = table->node(it);
                 auto  x = node.sigma.get(GRB_DoubleAttr_Xn);
-                if (x > 1e-4) {
+                if (x > EPS_CUT) {
                     if (it > 1) {
                         node.coeff_cut[0] += 1.0;
                         node.coeff_cut[1] += 1.0;
@@ -247,11 +246,11 @@ void ZeroHalfCuts::generate_cuts() {
                     }
 
                     for (int j = 0; j < 2; j++) {
-                        for (auto& it_aux : node.in_edges[j]) {
+                        for (auto& it_aux : node.in_edges.at(j)) {
                             auto aux = it_aux.lock();
                             if (aux) {
                                 auto& aux_node = table->node(*aux);
-                                aux_node.coeff_cut[j] -= 1.0;
+                                aux_node.coeff_cut.at(j) -= 1.0;
                             }
                         }
                     }
@@ -262,7 +261,7 @@ void ZeroHalfCuts::generate_cuts() {
                 for (auto& it : (*table)[i]) {
                     auto j = it.get_nb_job();
                     auto x = jobs_var[j].get(GRB_DoubleAttr_Xn);
-                    if (x > 1e-4) {
+                    if (x > EPS_CUT) {
                         it.coeff_cut[1] += 1.0;
                     }
                 }
@@ -313,15 +312,16 @@ void ZeroHalfCuts::generate_cuts() {
             //         dfs_lift(node.branch[0]);
             //     }
             // }
-            if (add)
+            if (add) {
                 construct_cut();
+            }
             // fmt::print("---------------------------------------------\n");
         }
         // getchar();
 
     } catch (GRBException& e) {
-        fmt::print("Error code = {0} {1:-^{2}} {3}", e.getErrorCode(), "", 40,
-                   e.getMessage());
+        fmt::print("Error code = {0} {1:-^{2}} {3}", e.getErrorCode(), "",
+                   ALIGN, e.getMessage());
     }
 }
 
@@ -329,11 +329,11 @@ void ZeroHalfCuts::dfs(const NodeId& v) {
     auto& node = table->node(v);
     node.lp_visited = true;
 
-    for (int i = 0; i < 2; i++) {
-        if (node.lp_x[i] > 1e-6) {
-            auto& child = table->node(node.branch[i]);
+    for (size_t i = 0U; i < 2; i++) {
+        if (node.lp_x.at(i) > EPS_CUT) {
+            auto& child = table->node(node.branch.at(i));
             if (!child.lp_visited) {
-                if (node.branch[i] == 1) {
+                if (node.branch.at(i) == 1) {
                     child.sigma = model->addVar(0.0, 1.0, 0.0, 'B',
                                                 fmt::format("sigma_terminal"));
                 } else {
@@ -343,23 +343,23 @@ void ZeroHalfCuts::dfs(const NodeId& v) {
                                     child.get_weight()));
                 }
 
-                node_ids.push_back(node.branch[i]);
+                node_ids.push_back(node.branch.at(i));
                 auto& s_source = node.sigma;
                 auto& s_head = child.sigma;
                 auto  str_y = fmt::format("y_{}_{}", node.get_nb_job(),
                                          node.get_weight());
                 auto  str_r = fmt::format("r_{}_{}", node.get_nb_job(),
                                          node.get_weight());
-                auto& y = node.y[i] =
-                    model->addVar(0.0, GRB_INFINITY, node.lp_x[i], 'B', str_y);
-                auto& r = node.r[i] =
+                auto& y = node.y.at(i) = model->addVar(
+                    0.0, GRB_INFINITY, node.lp_x.at(i), 'B', str_y);
+                auto& r = node.r.at(i) =
                     model->addVar(0.0, GRB_INFINITY, 0.0, 'I', str_r);
-                GRBLinExpr expr = -s_head + s_source - y - 2.0 * r;
+                GRBLinExpr expr = -s_head + s_source - y - HALF * r;
                 if (i) {
                     expr += jobs_var[node.get_nb_job()];
                 }
                 model->addConstr(expr, '=', 0.0);
-                dfs(node.branch[i]);
+                dfs(node.branch.at(i));
             } else {
                 auto& s_source = node.sigma;
                 auto& s_head = child.sigma;
@@ -367,11 +367,11 @@ void ZeroHalfCuts::dfs(const NodeId& v) {
                                          node.get_weight());
                 auto  str_r = fmt::format("r_{}_{}", node.get_nb_job(),
                                          node.get_weight());
-                auto& y = node.y[i] =
-                    model->addVar(0.0, GRB_INFINITY, node.lp_x[i], 'B', str_y);
-                auto& r = node.r[i] =
+                auto& y = node.y.at(i) = model->addVar(
+                    0.0, GRB_INFINITY, (node.lp_x)[i], 'B', str_y);
+                auto& r = node.r.at(i) =
                     model->addVar(0.0, GRB_INFINITY, 0.0, 'I', str_r);
-                GRBLinExpr expr = -s_head + s_source - y - 2.0 * r;
+                GRBLinExpr expr = -s_head + s_source - y - HALF * r;
                 if (i) {
                     expr += jobs_var[node.get_nb_job()];
                 }
@@ -386,26 +386,26 @@ void ZeroHalfCuts::dfs_lift(const NodeId& v) {
     auto& node = table->node(v);
 
     for (int k = 0; k < 2; k++) {
-        if (node.branch[k] <= 1) {
+        if (node.branch.at(k) <= 1) {
             continue;
         }
-        auto& child_node = table->node(node.branch[k]);
+        auto& child_node = table->node(node.branch.at(k));
 
-        if (node.lp_x[k] == 0.0 && !child_node.lp_visited) {
+        if (node.lp_x.at(k) == 0.0 && !child_node.lp_visited) {
             child_node.coeff_cut[0] += 1.0;
             child_node.coeff_cut[1] += 1.0;
 
             for (int j = 0; j < 2; j++) {
-                for (auto& it : child_node.in_edges[j]) {
+                for (auto& it : child_node.in_edges.at(j)) {
                     auto aux = it.lock();
                     if (aux) {
                         auto& aux_node = table->node(*aux);
-                        aux_node.coeff_cut[j] -= 1.0;
+                        aux_node.coeff_cut.at(j) -= 1.0;
                     }
                 }
             }
-            node_ids_lift.push_back(node.branch[k]);
-            dfs_lift(node.branch[k]);
+            node_ids_lift.push_back(node.branch.at(k));
+            dfs_lift(node.branch.at(k));
         }
     }
 }
